### 2024-06-14 #202 @rich-iannone
Great Tables submission
Submitting Author: Name (@rich-iannone)
All current maintainers: (@rich-iannone, @machow)
Package Name: Great Tables
One-Line Description of Package: Make awesome display tables using Python.
Repository Link:  https://github.com/posit-dev/great-tables
Version submitted: v0.9.0
EiC: TBD
Version accepted: TBD

---

The Great Tables package is all about creating tables for the purpose of presentation. You can use
Pandas or Polars DataFrames as inputs, and the Great Tables API allows you to:

- structure the data using column spanners and row groups, and add header and footer information
- format the data with a wide range of powerful formatting methods
- style the table to make it aesthetically pleasing or to highlight important information
- integrate the table display into notebooks, Quarto documents or web pages, and export the table
as HTML or a variety of image formats

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The package can be seen as a data visualization package, but it is perhaps more in the direction of data presentation/publication (i.e., not datavis in the traditional sense). However, tables are important and they are ubiquitous in all sorts of scientific publications.

The target audience is anyone who needs to present data in the tabular format. There is a particular focus on science and engineering applications as many of the formatting methods are geared toward this audience (e.g., scientific notation, significant figures, units notation, chemistry notation, etc.).

There are only a few packages that deal with tabular data presentation. The Pandas styler API is probably the best known of these, but it is limited in its capabilities. A big part of Great Tables is the ability to structure a table to a more traditional table format (one you'd commonly see in journals or reports) instead of interactive tables that are more common in web apps (i.e., displaying hundreds or thousands of rows of data). The formatting capabilities of Great Tables are also much more extensive than Pandas styler or other packages.

https://github.com/pyOpenSci/software-submission/issues/184
