### 2024-06-17 0680d5c
feat: GT.show() method for displaying tables (in browser)
