### 2024-06-17 6f9ce58
docs: add GT.show() to API docs
