+++ docs/_quarto.yml
@@ -58,6 +58,7 @@ filters:
   - interlinks
 
 interlinks:
+  fast: true
   sources:
     numpy:
       url: https://numpy.org/doc/stable/
@@ -182,6 +183,7 @@ quartodoc:
         also get the table code as an HTML fragment with the `as_raw_html()` method.
       contents:
         - GT.save
+        - GT.show
         - GT.as_raw_html
     - title: Value formatting functions
       desc: >
