### 2024-06-18 5a07ea1
Add documentation example
