### 2024-06-18 1f3af60
Add `column_units` and `column_pattern` attrs to ColInfo
