### 2024-06-18 4267bc9
Update comment in ColInfo class
