+++ great_tables/_gt_data.py
@@ -214,6 +214,8 @@ class ColInfo:
     # The components of the boxhead are:
     # `var` (obtained from column names)
     # `column_label` (obtained from column names)
+    # `column_units` (added through .cols_units())
+    # `column_pattern` (added through .cols_units())
     # `column_align` = None
     # `column_width` = None
 
