### 2024-06-18 f8e69ed
Add the `_set_column_units()` util fn
