### 2024-06-18 c243431
Add `replace_column_label()` method for ColInfo
