### 2024-06-18 c7680ed
Add `_boxhead_build()` method for GT class
