### 2024-06-21 9282372
Remove `_boxhead_build()`
