### 2024-06-21 94bceed
Subclass `Text` to add `Md` and `Html` classes
