### 2024-06-21 32998e8
Update `_process_text()` to handle more text classes
