### 2024-06-21 43ec68a
Handle units-syntax'd text in `cols_label()`
