### 2024-06-21 bb30863
Only convert to UniStr if text indicates units
