### 2024-06-22 #202 @Batalex
Hi there! Thank you for submitting your package for pyOpenSci review. Below are the basic checks that your package needs to pass to begin our review. If some of these are missing, we will ask you to work on them before the review process begins.

---
---

This is a really solid submission, and I am super excited for us to contribute to `great_tables`. I'll get started on finding an editor for this submission.

**suggestion**: TIL we can store CONTRIBUTING.md in `.github`. Would you mind adding a section in the README pointing to this file? I knew what I was looking for, but I think the entry barrier for new contributors would be lower if they did not have to open a PR to see the contributing guidelines.

**todo**: We can get started, but during the review's course, we should add the `paper.md` file as part of the review.
