### 2024-06-25 3e8e885
Remove unused `column_labels_units_pattern=` arg
