### 2024-06-25 0e79395
Remove `_set_column_units()` as it's unused
