### 2024-06-25 3075ded
Remove unneeded OptionsInfo entry
