### 2024-06-27 be94aa5
Add tests of class construction for Unit* classes
