### 2024-06-27 c31ff66
Add docs to cols_label() regarding units notation
