### 2024-06-27 f7fe652
docs: update cols_label docstring
