### 2024-06-27 b0dd277
Add example for cols_label() with unit not'n
