### 2024-06-28 9083a56
Better handle weight as str, int, or float
