### 2024-06-28 41b2a0f
Simplify setting of `table_font_names=`
