### 2024-06-28 8c9d91b
Remove redundant statement
