### 2024-06-28 11ddbb6
Remove validation of numerical font weight val
