### 2024-07-08 18efbc0
Add logic to allow units not'n in spanner labels
