### 2024-07-08 88874d5
Update type annotations in SpannerInfo
