+++ great_tables/_gt_data.py
@@ -759,7 +759,7 @@ class GroupRows(_Sequence[GroupRowInfo]):
 class SpannerInfo:
     spanner_id: str
     spanner_level: int
-    spanner_label: str | None = None
+    spanner_label: str | Md | Html | UnitStr | None = None
     spanner_units: str | None = None
     spanner_pattern: str | None = None
     vars: list[str] = field(default_factory=list)
