### 2024-07-08 8f5f2a8
Raise ValueError if input label not of specific types
