### 2024-07-08 48a3a9a
docs: draft blog for scientific publishing
