### 2024-07-08 fa64041
Update type annotations in SpannerInfo
