### 2024-07-08 4699b29
Tweak intro; add SciPy talk info
