### 2024-07-08 5a62917
fix: add more specific ModuleNotFoundError
