+++ great_tables/_tbl_data.py
@@ -330,7 +330,7 @@ def _(data: PlDataFrame, expr: Union[list[str], _selector_proxy_], strict: bool 
     pl_version = None
     try:
         pl_version = _re_version(version("polars"))
-    except:
+    except ModuleNotFoundError:
         pl_version = _re_version(version("polars-u64-idx"))
 
     expand_opts = {"strict": False} if pl_version >= (0, 20, 30) else {}
