### 2024-07-08 53dd27f
Add intro to 'Unit and scientific notation' section
