+++ docs/blog/tables-for-scientific-publishing/index.qmd
@@ -55,13 +55,13 @@ reactions_mini = (
         )
     )
 )
+
 (
     GT(reactions_mini, rowname_col="cmpd_name")
-    .tab_header(title="Gas-phase reactions \
-    of selected mercaptan compounds")
+    .tab_header(title="Gas-phase reactions of selected mercaptan compounds")
     .tab_spanner(
         columns=ps.ends_with("k298"),
-        label = "Reaction Rate Constant (298 K),<br>{{cm^3 molecules^–1 s^–1}}"
+        label="Reaction Rate Constant (298 K),<br>{{cm^3 molecules^–1 s^–1}}"
     )
     .fmt_units(columns="cmpd_formula")
     .fmt_scientific(columns=ps.ends_with("k298"))
@@ -81,11 +81,13 @@ reactions_mini = (
 )
 ```
 
-Note these pieces:
+This is a nice-looking table! And note these pieces:
 
-* The `label=` argument to functions like `tab_spanner()` supports a curly braces for special units syntax. `"{{cm^3 molecules^–1 s^–1}}"` becomes cm<sup>3</sup> TODO finish
-* The `fmt_units()` converts values that are unit syntax directly. For example the cell "%CH4S%"` becomes CH<sub>4</sub>S.
+* The `label=` argument to functions like `.tab_spanner()` supports the use of curly braces (`{{`/`}}`) for the specialized units notation. So using `"{{cm^3 molecules^–1 s^–1}}"` in the input will become cm<sup>3</sup> molecules<sup>–1</sup> s<sup>–1</sup> in the output
+* The `.fmt_units()` method converts values that are already in units notation in the table body. For example, a cell with text `"%CH4S%"` becomes CH<sub>4</sub>S (the surrounding `%` indicates that the text should be interpreted as chemistry notation).
+* The `.fmt_scientific()` method formats values (in this case, very small values) to scientific notation (e.g., 3.50 × 10<sup>–11</sup>). Not doing so would make the table look very strange to a researcher that is familar with this sort of data.
 
+The combination of units notation (and chemistry notation, which is a part of that) really makes the presentation of this table complete and understandable to a practioner of the field. Great Tables supports the use of units notation in spanner labels (with `.tab_spanner()`) and also in column labels (with `.cols_labels()`). The column label 'NO<sub>3</sub>' was created with the latter method by supplying the text `"{{%NO3%}}"` as the column label for the `NO3_k298` column.
 
 ## Nanoplots
 
