### 2024-07-08 c962e49
Refine GT table code
