+++ docs/blog/tables-for-scientific-publishing/index.qmd
@@ -36,7 +36,7 @@ To make this table work well in a scientific reporting context, we need three pi
 **Great Tables** provides the necessary functionality for all three requirements. Here is a summary table that tabulates rate constants for mercaptan compounds undergoing reaction with OH, O<sub>3</sub>, and Cl:
 
 ```{python}
-from great_tables import GT, md, system_fonts
+from great_tables import GT
 from great_tables.data import reactions
 import polars as pl
 import polars.selectors as ps
@@ -75,9 +75,7 @@ reactions_mini = (
     )
     .opt_stylize(style=1, color="blue")
     .opt_horizontal_padding(scale=3)
-    .tab_options(
-        table_font_names=system_fonts("humanist"),
-    )
+    .opt_table_font(stack="humanist")
 )
 ```
 
@@ -93,18 +91,18 @@ The combination of units notation (and chemistry notation, which is a part of th
 
 
 
-
 ```{python}
-from great_tables import GT, md, nanoplot_options
+from great_tables import GT, nanoplot_options
 from great_tables.data import gibraltar
 import polars as pl
 import polars.selectors as ps
-from datetime import datetime
+
 def to_seconds(timestr):
-    seconds= 0
+    seconds = 0
     for part in timestr.split(':'):
-        seconds= seconds*60 + int(part, 10)
+        seconds = seconds*60 + int(part, 10)
     return seconds
+
 nano_opts = nanoplot_options(
     data_point_radius=4,
     data_point_stroke_width=4,
@@ -116,25 +114,17 @@ nano_opts = nanoplot_options(
     show_data_points=True,
     show_data_area=False,
   )
+
 gibraltar_mini = (
     pl.from_pandas(gibraltar)
     .filter(pl.col("date") <= "2023-05-10")
-    #.with_columns(time=pl.concat_str(pl.col("time") + ":00"))
-    #.with_columns(
-    #    pl.col("time").map_elements(
-    #        to_seconds,
-    #        return_dtype=int
-    #    )
-    #)
     .with_columns(pl.col("humidity") * 100)
     .select(["date", "temp", "humidity"])
     .group_by("date")
-    .agg(
-        pl.col("temp"),
-        pl.col("humidity"),
-    )
+    .agg(pl.col("temp"), pl.col("humidity"))
     .sort("date")
 )
+
 (
   GT(gibraltar_mini)
   .tab_header(
