### 2024-07-08 6d25275
Add text around nanoplots example
