### 2024-07-08 eae2c15
Add author and date metadata
