### 2024-07-08 b3c4186
Modify text after nanoplots table
