### 2024-07-08 e4b6fd2
Ensure that nanoplots are left aligned
