### 2024-07-08 3a904bc
Update docs/blog/tables-for-scientific-publishing/index.qmd
Co-authored-by: Michael Chow <mc_al_github@fastmail.com>
