### 2024-07-15 242a5d0
Update polars examples for deprecated arg
