### 2024-07-15 cede326
chore: remove unused code
