### 2024-07-15 c390c86
feat: Heading.is_empty method
