### 2024-07-15 c041458
feat: Options.__getattr__ for accessing values
