### 2024-07-15 #202 @cmarmo
Hi @rich-iannone, thanks again for submitting Great Tables to pyOpenSci!
I'm Chiara and I'm following up as editor in chief for the next three months.
I'm glad to announce that @Batalex is volounteering to serve as editor too for your submission.... Thanks a lot Alex!
