### 2024-07-16 1111030
Tiny typo fixes and improvement suggestion to code example
BEFORE change:
- The nanoplot in the second example included more columns than described in the source note (i.e. days 3 through 9 plus "norm_I" and "norm_u" columns, instead of just days 3 through 8).
AFTER change:
- The nanoplot includes days 3 through 9 only (without "norm_I" and "norm_u")
- The source note on table footer is adjusted accordingly

Additional tiny typos:
- Remove unused "import great_tables"
- Remove superfluous "the" in text paragraph (line 142)
