### 2024-07-18 f1df81e
Tests row selection used for scaling nanoplots
