### 2024-07-27 b56073d
Consolidate ordered list code using new `_create_ordered_list()` function
