### 2024-07-27 e5009bf
Mention support for Polars in the `get-started` section
