### 2024-07-27 posit-dev/great-tables#409 @buhtz
docu: Mention the output format in README.md and Intro Article
I read the README and the Intro article.

I used the example code from the intro article. It run without any errors but I got no output. I just see the REPL shell `>>>` blinking.

Is this code meant to run in a jupyter notebook? If this is the case you should explicit mention that. Not all data-people working with things like that.

Is there an alternative for the rest of the Python users to use your package for tables? What is the output format? Vanilla java-script-free HTLM5, SVG, PNG, ...?
