### 2024-07-28 posit-dev/great-tables#409 @jrycw
@buhtz , thanks for your feedback. I've opened a discussion #410 with the team to address this issue. In the meantime, you might find it helpful to refer to a similar issue in #406 for some ideas on how to resolve it.
