### 2024-08-23 374f085
Make tag adjustments to heading component
