### 2024-08-23 365f2ab
Insert <thead> tags when assembling HTML parts
