### 2024-08-26 13d16ed
Restore line break on initial `<tr>` tag
