+++ great_tables/_utils_render_html.py
@@ -47,7 +47,8 @@ def create_heading_component_h(data: GTData) -> StringBuilder:
     <td colspan="{n_cols_total}" class="gt_heading gt_subtitle gt_font_normal gt_bottom_border">{subtitle}</td>
   </tr>"""
     else:
-        heading = f"""<tr class="gt_heading">
+        heading = f"""
+  <tr class="gt_heading">
     <td colspan="{n_cols_total}" class="gt_heading gt_title gt_font_normal">{title}</td>
   </tr>"""
 
