### 2024-08-26 00885d1
Remove Stringbuilder cls and usage
