### 2024-08-26 3d65e0b
docs: tweak polars example in intro
