+++ great_tables/__init__.py
@@ -19,6 +19,7 @@ from ._helpers import (
     pct,
     md,
     html,
+    google_font,
     random_id,
     system_fonts,
     define_units,
@@ -35,6 +36,7 @@ __all__ = (
     "pct",
     "md",
     "html",
+    "google_font",
     "system_fonts",
     "define_units",
     "nanoplot_options",
