### 2024-08-27 bd35847
Enable use of `table_additional_css` option
