### 2024-08-27 0b9418e
Add docs for `table_additional_css`
