### 2024-08-27 22997da
Store data for Google font in dict (in opt_table_font())
