### 2024-08-27 bac743e
Refactor conditional statements
