### 2024-08-28 148322c
Ensure `table_additional_css` init'd as a list
