### 2024-08-28 72db8cb
Create dataclass GoogleFont to work w/ helper
