### 2024-08-28 adc875d
Detect and use GoogleFont class + methods
