### 2024-08-28 b1802a4
Handle `additional_css` as a list of strings
