### 2024-08-28 8739dcb
Allow lists with mix of local and Google fonts
