### 2024-08-28 dc54ba4
Update docs for `google_font()`
