### 2024-08-29 d3562a4
Make adjustments to `google_font()` docs
