### 2024-08-29 036df24
Add tests for `google_font()` in two methods
