+++ tests/test_tab_create_modify.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import pytest
-from great_tables import GT
+from great_tables import GT, style, loc, google_font
 from great_tables._locations import LocBody
 from great_tables._styles import CellStyleFill
 from great_tables._tab_create_modify import tab_style
@@ -30,3 +30,19 @@ def test_tab_style_multiple_columns(gt: GT):
 
     assert len(new_gt._styles[0].styles) == 1
     assert new_gt._styles[0].styles[0] is style
+
+
+def test_tab_style_google_font(gt: GT):
+
+    new_gt = tab_style(
+        gt,
+        style=style.text(font=google_font(name="IBM Plex Mono")),
+        locations=loc.body(columns="time"),
+    )
+
+    rendered_html = new_gt.as_raw_html()
+
+    assert rendered_html.find(
+        "@import url('https://fonts.googleapis.com/css2?family=IBM+Plex+Mono&display=swap');"
+    )
+    assert rendered_html.find("font-family: IBM Plex Mono;")
