### 2024-08-30 600480c
Create OrderedSet class
