### 2024-08-30 2d45596
fix: up crud factor in .save method
