### 2024-08-30 e2238d0
fix: add small sleep after resizing to .save
