### 2024-08-30 0d0a4aa
Update documentation for `tab_spanner()`
