### 2024-08-30 ae9624b
fix: make debug borders less wide
