### 2024-09-05 9174182
Standardize the `GTSelf` signature
