+++ great_tables/_spanners.py
@@ -17,7 +17,7 @@ SpannerMatrix = "list[dict[str, str | None]]"
 
 
 def tab_spanner(
-    data: GTSelf,
+    self: GTSelf,
     label: str | Text,
     columns: SelectExpr = None,
     spanners: str | list[str] | None = None,
@@ -125,7 +125,7 @@ def tab_spanner(
     """
     from great_tables._helpers import UnitStr
 
-    crnt_spanner_ids = set([span.spanner_id for span in data._spanners])
+    crnt_spanner_ids = set([span.spanner_id for span in self._spanners])
 
     if id is None:
         # The label may contain HTML or Markdown, so we need to extract
@@ -153,7 +153,7 @@ def tab_spanner(
 
     # select columns ----
 
-    selected_column_names = resolve_cols_c(data=data, expr=columns, null_means="nothing") or []
+    selected_column_names = resolve_cols_c(data=self, expr=columns, null_means="nothing") or []
 
     # select spanner ids ----
     # TODO: this supports tidyselect
@@ -170,7 +170,7 @@ def tab_spanner(
         raise NotImplementedError("columns/spanners must be specified")
 
     # get column names associated with selected spanners ----
-    _vars = [span.vars for span in data._spanners if span.spanner_id in spanner_ids]
+    _vars = [span.vars for span in self._spanners if span.spanner_id in spanner_ids]
     spanner_column_names = list({k: True for k in itertools.chain(*_vars)})
 
     column_names = list({k: True for k in [*selected_column_names, *spanner_column_names]})
@@ -179,7 +179,7 @@ def tab_spanner(
 
     # get spanner level ----
     if level is None:
-        level = data._spanners.next_level(column_names)
+        level = self._spanners.next_level(column_names)
 
     # get spanner units and labels ----
     # TODO: grep units from {{.*}}, may need to switch delimiters
@@ -213,8 +213,8 @@ def tab_spanner(
         spanner_label=new_label,
     )
 
-    spanners = data._spanners.append_entry(new_span)
-    new_data = data._replace(_spanners=spanners)
+    spanners = self._spanners.append_entry(new_span)
+    new_data = self._replace(_spanners=spanners)
 
     if gather and not len(spanner_ids) and level == 0 and column_names:
         return cols_move(new_data, columns=column_names, after=column_names[0])
@@ -229,7 +229,7 @@ def _validate_sel_cols(sel_cols: list[str], col_vars: list[str]) -> None:
         raise ValueError("All `columns` must exist and be visible in the input `data` table.")
 
 
-def cols_move(data: GTSelf, columns: SelectExpr, after: str) -> GTSelf:
+def cols_move(self: GTSelf, columns: SelectExpr, after: str) -> GTSelf:
     """Move one or more columns.
 
     On those occasions where you need to move columns this way or that way, we can make use of the
@@ -287,11 +287,11 @@ def cols_move(data: GTSelf, columns: SelectExpr, after: str) -> GTSelf:
     if isinstance(columns, str):
         columns = [columns]
 
-    sel_cols = resolve_cols_c(data=data, expr=columns)
+    sel_cols = resolve_cols_c(data=self, expr=columns)
 
-    sel_after = resolve_cols_c(data=data, expr=[after])
+    sel_after = resolve_cols_c(data=self, expr=[after])
 
-    col_vars = [col.var for col in data._boxhead]
+    col_vars = [col.var for col in self._boxhead]
 
     if not len(sel_after):
[…]
