### 2024-09-09 169b5dd
refactor: use OrderedSet directly; rm _unique_set function
