+++ great_tables/_tbl_data.py
@@ -323,7 +323,7 @@ def _(
 def _(data: PlDataFrame, expr: Union[list[str], _selector_proxy_], strict: bool = True) -> _NamePos:
     # TODO: how to annotate type of a polars selector?
     # Seems to be polars.selectors._selector_proxy_.
-    from ._utils import _create_ordered_list
+    from ._utils import OrderedSet
     import polars.selectors as cs
 
     from polars import Expr
@@ -356,11 +356,11 @@ def _(data: PlDataFrame, expr: Union[list[str], _selector_proxy_], strict: bool 
         # this should be equivalent to reducing selectors using an "or" operator,
         # which isn't possible when there are selectors mixed with expressions
         # like pl.col("some_col")
-        final_columns = _create_ordered_list(
+        final_columns = OrderedSet(
             col_name
             for sel in all_selectors
             for col_name in cs.expand_selector(data, sel, **expand_opts)
-        )
+        ).as_list()
     else:
         if not isinstance(expr, (cls_selector, Expr)):
             raise TypeError(f"Unsupported selection expr type: {type(expr)}")
