### 2024-09-11 02ec399
Refactor `GT.opt_all_caps()`
