### 2024-09-12 f0ae1a7
Modify the `locations` parameter of `opt_all_caps()` to accept `Loc` objects
