+++ great_tables/loc.py
@@ -4,6 +4,7 @@ from ._locations import (
     LocBody as body,
     LocStub as stub,
     LocColumnLabels as column_labels,
+    LocRowGroups as row_group,
 )
 
-__all__ = ("body", "stub", "column_labels")
+__all__ = ("body", "stub", "column_labels", "row_group")
