### 2024-09-12 059275c
Update `opt_vertical_padding()` and `opt_horizontal_padding()`
