+++ great_tables/_options.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass, fields, replace
 from typing import TYPE_CHECKING, ClassVar, cast
 
 from great_tables import _utils
-from great_tables._helpers import FontStackName
+from great_tables._helpers import FontStackName, _intify_scaled_px, px
 
 
 if TYPE_CHECKING:
@@ -765,9 +765,7 @@ def opt_vertical_padding(self: GTSelf, scale: float = 1.0) -> GTSelf:
     # then reattach the units after the multiplication
     # TODO: a current limitation is that the padding values must be in pixels and not percentages
     # TODO: another limitation is that the returned values must be in integer pixel values
-    new_vertical_padding_vals = [
-        str(int(float(v.split("px")[0]) * scale)) + "px" for v in vertical_padding_vals
-    ]
+    new_vertical_padding_vals = [px(_intify_scaled_px(v, scale)) for v in vertical_padding_vals]
 
     return tab_options(self, **dict(zip(vertical_padding_params, new_vertical_padding_vals)))
 
@@ -864,9 +862,7 @@ def opt_horizontal_padding(self: GTSelf, scale: float = 1.0) -> GTSelf:
     # then reattach the units after the multiplication
     # TODO: a current limitation is that the padding values must be in pixels and not percentages
     # TODO: another limitation is that the returned values must be in integer pixel values
-    new_horizontal_padding_vals = [
-        str(int(float(v.split("px")[0]) * scale)) + "px" for v in horizontal_padding_vals
-    ]
+    new_horizontal_padding_vals = [px(_intify_scaled_px(v, scale)) for v in horizontal_padding_vals]
 
     return tab_options(self, **dict(zip(horizontal_padding_params, new_horizontal_padding_vals)))
 
