### 2024-09-12 83e5ae3
Show tables with style and color values applied
