### 2024-09-12 73f6e8c
Simplify docs in .tab_spanner()
