### 2024-09-12 37772be
Update _spanners.py
