+++ great_tables/_spanners.py
@@ -46,7 +46,7 @@ def tab_spanner(
         The text to use for the spanner label. We can optionally use the [`md()`](`great_tables.md`)
         and [`html()`](`great_tables.html`) helper functions to style the text as Markdown or to
         retain HTML elements in the text. Alternatively, units notation can be used (see
-        [`define_units()`](`great_tables.define_units`) for details.)
+        [`define_units()`](`great_tables.define_units`) for details).
     columns
         The columns to target. Can either be a single column name or a series of column names
         provided in a list.
