### 2024-09-12 1382d16
Refine docs for .opt_stylize()
