### 2024-09-12 8721eae
Link to opt_stylize() docs page from guide
