### 2024-09-12 2b0b2d4
Link to correct page in Get Started
