### 2024-09-16 97dba1c
Make updates to PyCon blog post
