### 2024-09-16 d5489ff
Update SciPy 2024 talk information
