### 2024-09-16 5b06a22
Make tweak to conference information
