### 2024-09-17 3a891dc
Add an example demonstrating `define_units()` usage
