### 2024-09-17 2f19948
Add `GT.cols_label()`
