### 2024-09-17 2ba634b
Make updates to PyCon 2024 post
