+++ docs/blog/pycon-2024-great-tables-are-possible/index.qmd
@@ -42,31 +42,25 @@ Note three important pieces:
 3. **The nanoplot** on the right shows a tiny bargraph for monthly sales over the past year.  This makes
    it easy to spot trends, and can be hovered over to get exact values.
 
-Critically, the code for this table used the DataFrame library [Polars](https://pola.rs/), which
-makes it really [easy to select rows and columns for styling](../polars-styling).
+Critically, the code for this table used the DataFrame library [Polars](https://pola.rs/), which makes it really [easy to select rows and columns for styling](../polars-styling/index.qmd).
 
 
 ## What's next?
 
 ### The 2024 Table Contest
 
-The world's premier display table contest---the [4th annual Table Contest](https://posit.co/blog/announcing-the-2024-table-contest/) draws competitors from near and far,
-to showcase the latest and greatest examples in table presentation.
-The contest is happening now, with **submissions due by May 31st, 2024**.
+The world's premier display table contest---the [4th annual Table Contest](https://posit.co/blog/announcing-the-2024-table-contest/) draws competitors from near and far, to showcase the latest and greatest examples in table presentation.
 
-For inspiration, see these resources:
-
-* [Contest announcement](https://posit.co/blog/announcing-the-2024-table-contest/)
-* [2022 winners and honorable mentions](https://posit.co/blog/winners-of-the-2022-table-contest/)
+The contest was a great success! On July 1, 2024 we announced the [2024 winners and honorable mentions](https://posit.co/blog/2024-table-contest-winners/). Check 'em out!
 
 
 ### posit::conf() workshop
 
-We're planning a posit::conf() 2024 workshop in August, called [Making Tables with gt and Great Tables](https://reg.conf.posit.co/flow/posit/positconf24/publiccatalog/page/publiccatalog/session/1707334049004001S0l2).
+We held a posit::conf() 2024 workshop, and you can find the materials at the [Making Tables with gt and Great Tables repo](https://github.com/posit-conf-2024/tables).
 
-If you're curious about making beautiful, publication quality tables in Python or R, we'd love to have you!
+If you're curious about making beautiful, publication quality tables in Python or R, do have a look at the resources at that GitHub repository.
 
-We’ll cover the following:
+We covered the following table topics:
 
 * Create table components and put them together (e.g., header, footer, stub, etc.)
 * Format cell values (numeric/scientific, date/datetime, etc.)
@@ -79,13 +73,11 @@ We’ll cover the following:
 
 Check out these resources to learn more about the wild and beautiful life of display tables:
 
-* [Great Tables example gallery](../examples/index.html)
-* [The Design Philosophy of Great Tables (blog post)](blog/design-philosophy/)
+* [Great Tables example gallery](../../examples/index.qmd)
+* [The Design Philosophy of Great Tables (blog post)](../design-philosophy/index.qmd)
 * [20 Minute Table Tutorial by Albert Rapp](https://youtu.be/ESyWcOFuMQc?si=1_bBRZEKENFKVNpB)
 * [PyCon talk: Making Beautiful, Publication Quality Tables is Possible in 2024](https://youtu.be/08yLWPpFdo4?si=vBK9h-ObXNKp9tHH)
 
 ## Hope all your tables are great!
 
-A huge thanks to all the people who have contributed to Great Tables over the past year.
-It's been a really incredible journey.
-Hope to see you at PyCon 2024!
+A huge thanks to all the people who have contributed to Great Tables over the past year. It's been a really incredible journey!
