### 2024-09-17 a7f5b4f
Use `OrderedSet.as_list()` to get unique CSS rules
