### 2024-09-17 dc859f0
Allow `font` to take a GoogleFont input
