### 2024-09-17 a043032
Add tests for various `cell_text(font=)` inputs
