### 2024-09-17 10d6cd9
refactor: prepare to use loc classes directly, not strings
