### 2024-09-17 39902a0
refactor: clean up last use of .locname
