+++ great_tables/_locations.py
@@ -473,11 +473,11 @@ def _(loc: LocHeader, data: GTData, style: list[CellStyle]) -> GTData:
 
     # set ----
     if loc.groups == "header":
-        info = StyleInfo(locname="header", locnum=1, styles=style)
+        info = StyleInfo(locname=loc, locnum=1, styles=style)
     elif loc.groups == "title":
-        info = StyleInfo(locname="title", locnum=2, styles=style)
+        info = StyleInfo(locname=loc, locnum=2, styles=style)
     elif loc.groups == "subtitle":
-        info = StyleInfo(locname="subtitle", locnum=3, styles=style)
+        info = StyleInfo(locname=loc, locnum=3, styles=style)
     else:
         raise ValueError(f"Unknown title group: {loc.groups}")
     return data._replace(_styles=data._styles + [info])
@@ -488,9 +488,7 @@ def _(loc: LocTitle, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(
-        _styles=data._styles + [StyleInfo(locname="title", locnum=1, styles=style)]
-    )
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
 
 
 @set_style.register
@@ -498,9 +496,7 @@ def _(loc: LocSubTitle, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(
-        _styles=data._styles + [StyleInfo(locname="subtitle", locnum=1, styles=style)]
-    )
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
 
 
 @set_style.register
@@ -508,9 +504,7 @@ def _(loc: LocStubhead, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(
-        _styles=data._styles + [StyleInfo(locname="stubhead", locnum=1, styles=style)]
-    )
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
 
 
 @set_style.register
@@ -518,9 +512,7 @@ def _(loc: LocStubheadLabel, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(
-        _styles=data._styles + [StyleInfo(locname="stubhead_label", locnum=1, styles=style)]
-    )
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
 
 
 @set_style.register
@@ -528,9 +520,7 @@ def _(loc: LocColumnLabels, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(
-        _styles=data._styles + [StyleInfo(locname="column_labels", locnum=1, styles=style)]
-    )
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
 
 
 @set_style.register
@@ -543,7 +533,7 @@ def _(loc: LocColumnLabel, data: GTData, style: list[CellStyle]) -> GTData:
     all_info: list[StyleInfo] = []
     for col_pos in positions:
         crnt_info = StyleInfo(
-            locname="column_label",
+            locname=loc,
             locnum=2,
             colname=col_pos.colname,
[…]
