### 2024-09-17 e91bbaa
Add `Returns` section to the doc for `GT.fmt_image()` and `GT.fmt_nanoplot()`
