+++ great_tables/data/__init__.py
@@ -308,6 +308,17 @@ This is a dataset with 13,545 rows and 5 columns.
 - `year`: The year for the population estimate.
 - `population`: The population estimate, midway through the year.
 
+Preview
+-------
+```
+  country_name country_code_2 country_code_3  year  population
+0        Aruba             AW            ABW  1960       54608
+1        Aruba             AW            ABW  1961       55811
+2        Aruba             AW            ABW  1962       56682
+3        Aruba             AW            ABW  1963       57475
+4        Aruba             AW            ABW  1964       58178
+```
+
 Source
 ------
 <https://data.worldbank.org/indicator/SP.POP.TOTL>
@@ -342,6 +353,17 @@ This is a dataset with 816 rows and 4 columns.
 zenith angle is calculated.
 - `sza`: The solar zenith angle in degrees, where missing values indicate that sunrise hadn't yet
 occurred by the `tst` value.
+
+Preview
+-------
+```
+  latitude month   tst  sza
+0       20   jan  0400  NaN
+1       20   jan  0430  NaN
+2       20   jan  0500  NaN
+3       20   jan  0530  NaN
+4       20   jan  0600  NaN
+```
 
 Source
 ------
@@ -383,6 +405,25 @@ car could have automatic transmission (`"a"`), manual transmission (`"m"`), an o
 between both types (`"am"`), or, direct drive (`"dd"`)
 - `ctry_origin`: The country name for where the vehicle manufacturer is headquartered.
 - `msrp`: Manufacturer's suggested retail price in U.S. dollars (USD).
+
+Preview
+-------
+```
+       mfr         model    year        trim    bdy_style     hp  hp_rpm  \
+0     Ford            GT  2017.0  Base Coupe        coupe  647.0  6250.0
+1  Ferrari  458 Speciale  2015.0  Base Coupe        coupe  597.0  9000.0
+2  Ferrari    458 Spider  2015.0        Base  convertible  562.0  9000.0
+3  Ferrari    458 Italia  2014.0  Base Coupe        coupe  562.0  9000.0
+4  Ferrari       488 GTB  2016.0  Base Coupe        coupe  661.0  8000.0
+
+     trq  trq_rpm  mpg_c  mpg_h drivetrain trsmn    ctry_origin      msrp
+0  550.0   5900.0   11.0   18.0        rwd    7a  United States  447000.0
+1  398.0   6000.0   13.0   17.0        rwd    7a          Italy  291744.0
+2  398.0   6000.0   13.0   17.0        rwd    7a          Italy  263553.0
+3  398.0   6000.0   13.0   17.0        rwd    7a          Italy  233509.0
+4  561.0   3000.0   15.0   22.0        rwd    7a          Italy  245400.0
+```
+
 """
 
 sp500: pd.DataFrame = pd.read_csv(_sp500_fname, dtype=_sp500_dtype)  # type: ignore
@@ -402,6 +443,25 @@ This is a dataset with 16,607 rows and 7 columns.
 `close` price is adjusted for splits.
 - `volume`: The number of trades for the given `date`.
 - `adj_close`: The close price adjusted for both dividends and splits.
+
+Preview
+-------
+```
+         date       open       high      low      close        volume  \
+0  2015-12-31  2060.5901  2062.5400  2043.62  2043.9399  2.655330e+09
+1  2015-12-30  2077.3401  2077.3401  2061.97  2063.3601  2.367430e+09
+2  2015-12-29  2060.5400  2081.5601  2060.54  2078.3601  2.542000e+09
[…]
