### 2024-09-18 779361d
Update `GT.save()`
