### 2024-09-19 20fa355
Use `polars.DataFrame.glimpse` to generate dataset previews
