+++ great_tables/data/__init__.py
@@ -311,12 +311,13 @@ This is a dataset with 13,545 rows and 5 columns.
 Preview
 -------
 ```
-  country_name country_code_2 country_code_3  year  population
-0        Aruba             AW            ABW  1960       54608
-1        Aruba             AW            ABW  1961       55811
-2        Aruba             AW            ABW  1962       56682
-3        Aruba             AW            ABW  1963       57475
-4        Aruba             AW            ABW  1964       58178
+Rows: 13545
+Columns: 5
+$ country_name   <str> 'Aruba', 'Aruba', 'Aruba'
+$ country_code_2 <str> 'AW', 'AW', 'AW'
+$ country_code_3 <str> 'ABW', 'ABW', 'ABW'
+$ year           <i64> 1960, 1961, 1962
+$ population     <i64> 54608, 55811, 56682
 ```
 
 Source
@@ -357,12 +358,12 @@ occurred by the `tst` value.
 Preview
 -------
 ```
-  latitude month   tst  sza
-0       20   jan  0400  NaN
-1       20   jan  0430  NaN
-2       20   jan  0500  NaN
-3       20   jan  0530  NaN
-4       20   jan  0600  NaN
+Rows: 816
+Columns: 4
+$ latitude <str> '20', '20', '20'
+$ month    <str> 'jan', 'jan', 'jan'
+$ tst      <str> '0400', '0430', '0500'
+$ sza      <f64> None, None, None
 ```
 
 Source
@@ -409,19 +410,23 @@ between both types (`"am"`), or, direct drive (`"dd"`)
 Preview
 -------
 ```
-       mfr         model    year        trim    bdy_style     hp  hp_rpm  \
-0     Ford            GT  2017.0  Base Coupe        coupe  647.0  6250.0
-1  Ferrari  458 Speciale  2015.0  Base Coupe        coupe  597.0  9000.0
-2  Ferrari    458 Spider  2015.0        Base  convertible  562.0  9000.0
-3  Ferrari    458 Italia  2014.0  Base Coupe        coupe  562.0  9000.0
-4  Ferrari       488 GTB  2016.0  Base Coupe        coupe  661.0  8000.0
-
-     trq  trq_rpm  mpg_c  mpg_h drivetrain trsmn    ctry_origin      msrp
-0  550.0   5900.0   11.0   18.0        rwd    7a  United States  447000.0
-1  398.0   6000.0   13.0   17.0        rwd    7a          Italy  291744.0
-2  398.0   6000.0   13.0   17.0        rwd    7a          Italy  263553.0
-3  398.0   6000.0   13.0   17.0        rwd    7a          Italy  233509.0
-4  561.0   3000.0   15.0   22.0        rwd    7a          Italy  245400.0
+Rows: 47
+Columns: 15
+$ mfr         <str> 'Ford', 'Ferrari', 'Ferrari'
+$ model       <str> 'GT', '458 Speciale', '458 Spider'
+$ year        <f64> 2017.0, 2015.0, 2015.0
+$ trim        <str> 'Base Coupe', 'Base Coupe', 'Base'
+$ bdy_style   <str> 'coupe', 'coupe', 'convertible'
+$ hp          <f64> 647.0, 597.0, 562.0
+$ hp_rpm      <f64> 6250.0, 9000.0, 9000.0
+$ trq         <f64> 550.0, 398.0, 398.0
+$ trq_rpm     <f64> 5900.0, 6000.0, 6000.0
+$ mpg_c       <f64> 11.0, 13.0, 13.0
+$ mpg_h       <f64> 18.0, 17.0, 17.0
+$ drivetrain  <str> 'rwd', 'rwd', 'rwd'
+$ trsmn       <str> '7a', '7a', '7a'
[…]
