### 2024-09-19 aa03028
Add tests for `val_fmt_image()`
