### 2024-09-19 5340b99
Strengthen assertion assumptions
