+++ tests/test_formats_vals.py
@@ -15,16 +15,17 @@ def test_locate_val_fmt_image(img_paths: Path):
     imgs = vals.fmt_image("1", path=img_paths, file_pattern="metro_{}.svg")
     with open(img_paths / "metro_1.svg", "rb") as f:
         encoded = base64.b64encode(f.read()).decode()
+
     assert encoded in imgs[0]
 
 
 def test_val_fmt_image_single(img_paths: Path):
     imgs = vals.fmt_image("1", path=img_paths, file_pattern="metro_{}.svg")
-    assert "data:image/svg+xml" in imgs[0]
+    assert 'img src="data:image/svg+xml;base64' in imgs[0]
 
 
 def test_val_fmt_image_multiple(img_paths: Path):
     img1, img2 = vals.fmt_image(["1", "2"], path=img_paths, file_pattern="metro_{}.svg")
 
-    assert "data:image/svg+xml;base64" in img1
-    assert "data:image/svg+xml;base64" in img2
+    assert 'img src="data:image/svg+xml;base64' in img1
+    assert 'img src="data:image/svg+xml;base64' in img2
