+++ docs/_quarto.yml
@@ -46,6 +46,7 @@ website:
             - get-started/column-selection.qmd
             - get-started/row-selection.qmd
             - get-started/nanoplots.qmd
+            - get-started/targeted-styles.qmd
 
 format:
   html:
