### 2024-09-19 eced055
tests: add kitchen sink location snapshot test
