### 2024-09-23 58436b1
fix: use full html page in show() for correct utf-8 display
