### 2024-09-24 f79e7e6
update `_save_screenshot()`
