### 2024-09-24 2fbe4da
Implement row striping in HTML table
