### 2024-09-24 0686857
Improve documentation for `opt_row_striping()`
