### 2024-09-24 6225204
Add example to `opt_row_striping()`
