### 2024-09-24 d113779
Add tests for row striping (including snapshot)
