### 2024-09-24 9d66bf7
Add `opt_row_striping()` to reference docs
