### 2024-09-24 e4b6b6c
fix: resolve_rows_i can always handle None expr
