### 2024-09-24 978bce4
feat: implement LocRowGroupLabel styles
