### 2024-09-24 2fab150
feat: support LocRowLabel styles
