### 2024-09-24 7cccdaf
feat: support LocSourceNotes styles
