### 2024-09-24 8f5dada
tests: add source note to styles test, update snapshot
