### 2024-09-24 b549c7e
tests: correctly target row of data in snapshot
