### 2024-09-24 8b4b811
refactor: remove LocStubheadLabel
