### 2024-09-24 a3bf5df
chore: remove print statement
