### 2024-09-25 e3e6134
Modify tests based on code review comments
