### 2024-09-25 561dd1e
Consolidate table body generation code
