### 2024-09-25 663fce8
Use separate snapshots for striping of stub & body
