### 2024-09-25 5548da1
feat: implement row striping options (#461)
* Implement row striping in HTML table

* Improve documentation for `opt_row_striping()`

* Add example to `opt_row_striping()`

* Add tests for row striping (including snapshot)

* Add `opt_row_striping()` to reference docs

* Modify tests based on code review comments

* Consolidate table body generation code

* Use separate snapshots for striping of stub & body
