### 2024-09-25 f84e87f
Add `add_row_striping` arg to `opt_stylize()`
