### 2024-09-25 3d8346b
Add description of new argument in docs
