### 2024-09-25 f41c7ad
Add `row_striping_background_color` to StyleMapper
