### 2024-09-25 8bf5966
Remove `table_outline_color` from styles dict
