### 2024-09-25 f0e47a8
Ensure row striping param is added to params dict
