### 2024-09-25 1620e3a
Add missing table border for styles 2, 4, and 5
