### 2024-09-26 778ed4e
Add tests of `opt_stylize()` w/ and w/o striping
