### 2024-09-26 0dbdcfa
Add test to verify border props in all style vals
