### 2024-09-27 ad546bd
Add WIP implementation of `fmt_duration()`
