+++ docs/_quarto.yml
@@ -114,6 +114,7 @@ quartodoc:
         - GT.fmt_date
         - GT.fmt_time
         - GT.fmt_datetime
+        - GT.fmt_duration
         - GT.fmt_markdown
         - GT.fmt_units
         - GT.fmt_image
