### 2024-09-27 29e3526
Include `google_font()` helper fn in API reference
