### 2024-09-27 3ba7818
feat: implement loc.headers
