### 2024-09-27 90095a4
feat: implement LocFooter styles
