### 2024-09-27 9e623e0
refactor: remove groups attr from Loc classes
