### 2024-09-27 1a2e9d4
refactor: remove StyleInfo.locnum attr, .loc always a Loc class
