+++ great_tables/_locations.py
@@ -17,14 +17,6 @@ from ._gt_data import (
     GTData,
     Spanners,
     StyleInfo,
-    LocName,
-    LocHeaderName,
-    LocStubheadName,
-    LocColumnLabelsName,
-    LocStubName,
-    LocBodyName,
-    LocFooterName,
-    LocUnknownName,
 )
 from ._styles import CellStyle
 from ._tbl_data import PlDataFrame, PlExpr, eval_select, eval_transform
@@ -461,7 +453,7 @@ def _(loc: LocHeader, data: GTData, style: list[CellStyle]) -> GTData:
     for entry in style:
         entry._raise_if_requires_data(loc)
 
-    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=3, styles=style)])
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, styles=style)])
 
 
 @set_style.register
@@ -469,7 +461,7 @@ def _(loc: LocTitle, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, styles=style)])
 
 
 @set_style.register
@@ -477,7 +469,7 @@ def _(loc: LocSubTitle, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, styles=style)])
 
 
 @set_style.register
@@ -485,7 +477,7 @@ def _(loc: LocStubhead, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, styles=style)])
 
 
 @set_style.register
@@ -493,7 +485,7 @@ def _(loc: LocStubheadLabel, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, styles=style)])
 
 
 @set_style.register
@@ -501,7 +493,7 @@ def _(loc: LocColumnLabels, data: GTData, style: list[CellStyle]) -> GTData:
     # validate ----
     for entry in style:
         entry._raise_if_requires_data(loc)
-    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, locnum=1, styles=style)])
+    return data._replace(_styles=data._styles + [StyleInfo(locname=loc, styles=style)])
 
 
 @set_style.register
@@ -515,7 +507,6 @@ def _(loc: LocColumnLabel, data: GTData, style: list[CellStyle]) -> GTData:
     for col_pos in positions:
         crnt_info = StyleInfo(
             locname=loc,
-            locnum=2,
             colname=col_pos.colname,
             rownum=col_pos.row,
             styles=styles,
@@ -533,8 +524,7 @@ def _(loc: LocSpannerLabel, data: GTData, style: list[CellStyle]) -> GTData:
 
     new_loc = resolve(loc, data._spanners)
[…]
