### 2024-09-27 e2bea6a
refactor: remove redundant or unimplemented set_style concretes
