### 2024-09-27 cbc39bd
refactor!: rename or remove locations based on Rich feedback
