### 2024-09-27 bdbe223
refactor!: rename loc.row_group_labels to row_groups
