### 2024-09-27 febbd70
docs: flesh out targeted styles a bit
