### 2024-09-27 b2da03c
Allow for granular section restyling via  convenience api (#341)
* WIP allow for granular section restyling via  convenience api

* fix lint

* working on tests

* refactor: prepare to use loc classes directly, not strings

* refactor: clean up last use of .locname

* fix: support spanner label targeting

* feat: add style.css for raw css

* tests: add kitchen sink location snapshot test

* tests: update snapshots

* fix: title style tag needs space before

* fix: resolve_rows_i can always handle None expr

* feat: implement LocRowGroupLabel styles

* feat: support LocRowLabel styles

* feat: 
[…]
