+++ great_tables/_utils_render_html.py
@@ -1,19 +1,48 @@
 from __future__ import annotations
 
-from itertools import chain
+from itertools import chain, groupby
+from math import isnan
 from typing import Any, cast
 
 from great_tables._spanners import spanners_print_matrix
 from htmltools import HTML, TagList, css, tags
 
-from ._gt_data import GTData
+from ._gt_data import GTData, Styles, GroupRowInfo
 from ._tbl_data import _get_cell, cast_frame_to_string, n_rows, replace_null_frame
 from ._text import _process_text, _process_text_id
 from ._utils import heading_has_subtitle, heading_has_title, seq_groups
+from . import _locations as loc
+
+
+def _is_loc(loc: str | loc.Loc, cls: type[loc.Loc]):
+    if isinstance(loc, str):
+        return loc == cls.groups
+
+    return isinstance(loc, cls)
+
+
+def _flatten_styles(styles: Styles, wrap: bool = False) -> str:
+    # flatten all StyleInfo.styles lists
+    style_entries = list(chain(*[x.styles for x in styles]))
+    rendered_styles = [el._to_html_style() for el in style_entries]
+
+    # TODO dedupe rendered styles in sequence
+
+    if wrap:
+        if rendered_styles:
+            # return style html attribute
+            return f' style="{" ".join(rendered_styles)}"'
+        # if no rendered styles, just return a blank
+        return ""
+    if rendered_styles:
+        # return space-separated list of rendered styles
+        return " ".join(rendered_styles)
+    # if not wrapping the styles for html element,
+    # return None so htmltools omits a style attribute
+    return None
 
 
 def create_heading_component_h(data: GTData) -> str:
-
     title = data._heading.title
     subtitle = data._heading.subtitle
 
@@ -31,6 +60,13 @@ def create_heading_component_h(data: GTData) -> str:
     title = _process_text(title)
     subtitle = _process_text(subtitle)
 
+    # Filter list of StyleInfo for the various header components
+    styles_header = [x for x in data._styles if _is_loc(x.locname, loc.LocHeader)]
+    styles_title = [x for x in data._styles if _is_loc(x.locname, loc.LocTitle)]
+    styles_subtitle = [x for x in data._styles if _is_loc(x.locname, loc.LocSubTitle)]
+    title_style = _flatten_styles(styles_header + styles_title, wrap=True)
+    subtitle_style = _flatten_styles(styles_header + styles_subtitle, wrap=True)
+
     # Get the effective number of columns, which is number of columns
     # that will finally be rendered accounting for the stub layout
     n_cols_total = data._boxhead._get_effective_number_of_columns(
@@ -40,15 +76,15 @@ def create_heading_component_h(data: GTData) -> str:
     if has_subtitle:
         heading = f"""
   <tr class="gt_heading">
-    <td colspan="{n_cols_total}" class="gt_heading gt_title gt_font_normal">{title}</td>
+    <td colspan="{n_cols_total}" class="gt_heading gt_title gt_font_normal"{title_style}>{title}</td>
   </tr>
   <tr class="gt_heading">
-    <td colspan="{n_cols_total}" class="gt_heading gt_subtitle gt_font_normal gt_bottom_border">{subtitle}</td>
+    <td colspan="{n_cols_total}" class="gt_heading gt_subtitle gt_font_normal gt_bottom_border"{subtitle_style}>{subtitle}</td>
   </tr>"""
     else:
         heading = f"""
   <tr class="gt_heading">
-    <td colspan="{n_cols_total}" class="gt_heading gt_title gt_font_normal">{title}</td>
+    <td colspan="{n_cols_total}" class="gt_heading gt_title gt_font_normal"{title_style}>{title}</td>
   </tr>"""
 
[…]
