+++ great_tables/_locations.py
@@ -1,16 +1,23 @@
 from __future__ import annotations
 
 import itertools
-from dataclasses import dataclass
+from dataclasses import dataclass, field
 from functools import singledispatch
-from typing import TYPE_CHECKING, Any, Callable, Literal
+from typing import TYPE_CHECKING, Any, Callable, ClassVar, Literal, Union
 
 from typing_extensions import TypeAlias
 
 # note that types like Spanners are only used in annotations for concretes of the
 # resolve generic, but we need to import at runtime, due to singledispatch looking
 # up annotations
-from ._gt_data import ColInfoTypeEnum, FootnoteInfo, FootnotePlacement, GTData, Spanners, StyleInfo
+from ._gt_data import (
+    ColInfoTypeEnum,
+    FootnoteInfo,
+    FootnotePlacement,
+    GTData,
+    Spanners,
+    StyleInfo,
+)
 from ._styles import CellStyle
 from ._tbl_data import PlDataFrame, PlExpr, eval_select, eval_transform
 
@@ -35,6 +42,7 @@ class CellPos:
     column: int
     row: int
     colname: str
+    rowname: str | None = None
 
 
 @dataclass
@@ -43,42 +51,62 @@ class Loc:
 
 
 @dataclass
-class LocTitle(Loc):
+class LocHeader(Loc):
     """A location for targeting the table title and subtitle."""
 
-    groups: Literal["title", "subtitle"]
+
+@dataclass
+class LocTitle(Loc):
+    """A location for targeting the title."""
+
+
+@dataclass
+class LocSubTitle(Loc):
+    """A location for targeting the subtitle."""
 
 
 @dataclass
 class LocStubhead(Loc):
-    groups: Literal["stubhead"] = "stubhead"
+    """A location for targeting the table stubhead and stubhead label."""
 
 
 @dataclass
-class LocColumnSpanners(Loc):
-    """A location for column spanners."""
+class LocStubheadLabel(Loc):
+    """A location for targetting the stubhead."""
 
-    # TODO: these can also be tidy selectors
-    ids: list[str]
+
+@dataclass
+class LocColumnHeader(Loc):
+    """A location for column spanners and column labels."""
 
 
 @dataclass
 class LocColumnLabels(Loc):
-    # TODO: these can be tidyselectors
-    columns: list[str]
+    columns: SelectExpr = None
 
 
 @dataclass
-class LocRowGroups(Loc):
-    # TODO: these can be tidyselectors
-    groups: list[str]
+class LocSpannerLabels(Loc):
+    """A location for column spanners."""
+
+    ids: SelectExpr = None
 
 
 @dataclass
 class LocStub(Loc):
-    # TODO: these can be tidyselectors
-    # TODO: can this take integers?
-    rows: list[str]
+    """A location for targeting the table stub, row group labels, summary labels, and body."""
+
+    rows: RowSelectExpr = None
+
+
+@dataclass
+class LocRowGroups(Loc):
+    rows: RowSelectExpr = None
+
+
+@dataclass
+class LocSummaryLabel(Loc):
+    rows: RowSelectExpr = None
 
 
 @dataclass
@@ -108,6 +136,7 @@ class LocBody(Loc):
     ------
     See [`GT.tab_style()`](`great_tables.GT.tab_style`).
     """
+
     columns: SelectExpr = None
     rows: RowSelectExpr = None
 
@@ -115,41 +144,23 @@ class LocBody(Loc):
 @dataclass
 class LocSummary(Loc):
     # TODO: these can be tidyselectors
-    groups: list[str]
-    columns: list[str]
-    rows: list[str]
+    columns: SelectExpr = None
+    rows: RowSelectExpr = None
 
 
 @dataclass
-class LocGrandSummary(Loc):
-    # TODO: these can be tidyselectors
[…]

[…]
