### 2024-09-27 8d485ee
Revert "Allow for granular section restyling via  convenience api (#341)"
This reverts commit b2da03cee5a9d3f2e311069750ef8f1bc25dad84.
