+++ great_tables/style.py
@@ -4,7 +4,6 @@ from ._styles import (
     CellStyleText as text,
     CellStyleFill as fill,
     CellStyleBorders as borders,
-    CellStyleCss as css,
 )
 
-__all__ = ("text", "fill", "borders", "css")
+__all__ = ("text", "fill", "borders")
