### 2024-09-28 f3c69c0
Update test for `UnitStr`
