### 2024-09-30 b0edc73
Add docstring for LocSourceNotes
