+++ great_tables/_locations.py
@@ -160,7 +160,40 @@ class LocFootnotes(Loc):
 
 @dataclass
 class LocSourceNotes(Loc):
-    """A location for targeting source notes."""
+    """A location specification for targeting the source notes.
+
+    The `loc.source_notes()` class is used to target the source notes in the table. The class can be
+    used to apply custom styling with the [`tab_style()`](`great_tables.GT.tab_style`) method. That
+    method has a `locations=` argument and this class should be used there to perform the targeting.
+    The 'source_notes' location is generated by
+    [`tab_source_note()`](`great_tables.GT.tab_source_note`).
+
+    Returns
+    -------
+    LocSourceNotes
+        A `LocSourceNotes` object, which is used for a `locations=` argument if specifying the
+        source notes.
+
+    Examples
+    --------
+    Let's use a subset of the [`gtcars`] dataset in a new table. Add a source note (with
+    [`tab_source_note()`](`great_tables.GT.tab_source_note`) and style the source notes section
+    inside [`tab_style()`](`great_tables.GT.tab_style`) with `locations=loc.source_notes()`.
+
+    ```{python}
+    from great_tables import GT, style, loc
+    from great_tables.data import gtcars
+
+    (
+        GT(gtcars[["mfr", "model", "msrp"]].head(5))
+        .tab_source_note(source_note="From edmunds.com")
+        .tab_style(
+            style=style.text(color="blue", size="small", weight="bold"),
+            locations=loc.source_notes()
+        )
+    )
+    ```
+    """
 
 
 # Utils ================================================================================
