### 2024-09-30 326e523
Add `loc.source_note` to API reference
