+++ docs/_quarto.yml
@@ -146,7 +146,7 @@ quartodoc:
         specification of the styling properties to be applied to the targeted locations.
       contents:
         - loc.body
-        - loc.source_note
+        - loc.source_notes
         - style.fill
         - style.text
         - style.borders
