### 2024-09-30 a203fbc
Revise text in some Loc* docstrings
