+++ great_tables/_locations.py
@@ -114,9 +114,9 @@ class LocBody(Loc):
     # TODO: these can be tidyselectors
     """A location specification for targeting data cells in the table body.
 
-    The `loc.body()` class is used to target the data cells in the table body. The class can be used
-    to apply custom styling with the `tab_style()` method. That method has a `locations` argument
-    and this class should be used there to perform the targeting.
+    With `loc.body()`, we can target the data cells in the table body. This is useful for applying
+    custom styling with the [`tab_style()`](`great_tables.GT.tab_style`) method. That method has a
+    `locations=` argument and this class should be used there to perform the targeting.
 
     Parameters
     ----------
@@ -130,7 +130,7 @@ class LocBody(Loc):
     Returns
     -------
     LocBody
-        A LocBody object, which is used for a `locations` argument if specifying the table body.
+        A LocBody object, which is used for a `locations=` argument if specifying the table body.
 
     Examples
     ------
@@ -150,7 +150,39 @@ class LocSummary(Loc):
 
 @dataclass
 class LocFooter(Loc):
-    """A location for targeting the footer."""
+    """A location specification for targeting the table footer.
+
+    With `loc.footer()` we can target the table's footer. This is useful when applying custom
+    styling with the [`tab_style()`](`great_tables.GT.tab_style`) method. That method has a
+    `locations=` argument and this class should be used there to perform the targeting. The 'footer'
+    location is generated by [`tab_source_note()`](`great_tables.GT.tab_source_note`).
+
+    Returns
+    -------
+    LocFooter
+        A `LocFooter` object, which is used for a `locations=` argument if specifying the footer of
+        the table.
+
+    Examples
+    --------
+    Let's use a subset of the [`gtcars`] dataset in a new table. Add a source note (with
+    [`tab_source_note()`](`great_tables.GT.tab_source_note`) and style this footer section inside of
+    [`tab_style()`](`great_tables.GT.tab_style`) with `locations=loc.footer()`.
+
+    ```{python}
+    from great_tables import GT, style, loc
+    from great_tables.data import gtcars
+
+    (
+        GT(gtcars[["mfr", "model", "msrp"]].head(5))
+        .tab_source_note(source_note="From edmunds.com")
+        .tab_style(
+            style=style.text(color="blue", size="small", weight="bold"),
+            locations=loc.footer()
+        )
+    )
+    ```
+    """
 
 
 @dataclass
@@ -162,10 +194,10 @@ class LocFootnotes(Loc):
 class LocSourceNotes(Loc):
     """A location specification for targeting the source notes.
 
-    The `loc.source_notes()` class is used to target the source notes in the table. The class can be
-    used to apply custom styling with the [`tab_style()`](`great_tables.GT.tab_style`) method. That
-    method has a `locations=` argument and this class should be used there to perform the targeting.
-    The 'source_notes' location is generated by
+    With `loc.source_notes()`, we can target the source notes in the table. This is useful when
+    applying custom with the [`tab_style()`](`great_tables.GT.tab_style`) method. That method has a
+    `locations=` argument and this class should be used there to perform the targeting. The
+    'source_notes' location is generated by
     [`tab_source_note()`](`great_tables.GT.tab_source_note`).
 
     Returns
