### 2024-09-30 6f7559f
Add docstring for LocColumnLabels
