### 2024-09-30 e5255a9
Add loc.footer to API reference
