### 2024-09-30 9a9c3d6
Make correction to 'Returns' section in docstring
