### 2024-09-30 c4c281f
Add docstring for the LocTitle class
