### 2024-09-30 617ac10
Add docstring for LocSubTitle class
