### 2024-09-30 b807c14
Add docstring for LocHeader class
