### 2024-09-30 1d5a414
docs: do not document GT members inline on its reference page
