### 2024-09-30 a407a98
docs: link to method docs inline on GT reference page
