### 2024-09-30 b89d94e
chore: format _quarto.yml
