### 2024-09-30 a453212
Add docstring for LocSpannerLabels class
