### 2024-09-30 d1a357c
Make adjustments to text introducing example
