### 2024-10-01 cc53a16
Add docstring for LocColumnHeader
