### 2024-10-01 164cea6
Add docstring for LocStubheadLabel
