### 2024-10-01 0e91b1d
Document LocStubhead instead of LocStubheadLabel
