### 2024-10-01 4f14979
Add docstring for LocStub class
