### 2024-10-01 8c82aef
Add docstring for LocRowGroups class
