### 2024-10-01 c1fcebe
Add example to LocBody docstring
