### 2024-10-01 606cf68
Reorder loc methods in API reference
