### 2024-10-01 236a6d2
Add note about future 'footnotes' loc in LocFooter
