+++ great_tables/_locations.py
@@ -552,7 +552,8 @@ class LocSummary(Loc):
 class LocFooter(Loc):
     """A location specification for targeting the table footer.
 
-    With `loc.footer()` we can target the table's footer. This is useful when applying custom
+    With `loc.footer()` we can target the table's footer, which currently contains the source notes
+    (and may contain a 'footnotes' location in the future). This is useful when applying custom
     styling with the [`tab_style()`](`great_tables.GT.tab_style`) method. That method has a
     `locations=` argument and this class should be used there to perform the targeting. The 'footer'
     location is generated by [`tab_source_note()`](`great_tables.GT.tab_source_note`).
