### 2024-10-01 6c1d6b2
Get components in place for LaTeX rendering
