### 2024-10-01 e9fc9e3
Tweak docstring titles for Loc* classes
