### 2024-10-02 f16e9c5
Add two `latex*` options in `tab_options()`
