### 2024-10-02 53477b5
Add the `check_quarto()` util function
