### 2024-10-02 9344487
Define logic in `create_wrap_start_l()`
