+++ great_tables/_export.py
@@ -177,13 +177,31 @@ WebDrivers: TypeAlias = Literal[
 DebugDumpOptions: TypeAlias = Literal["zoom", "width_resize", "final_resize"]
 
 
+class _NoOpDriverCtx:
+    """Context manager that no-ops entering a webdriver(options=...) instance."""
+
+    def __init__(self, driver: webdriver.Remote):
+        self.driver = driver
+
+    def __call__(self, options):
+        # no-op what is otherwise instantiating webdriver with options,
+        # since a webdriver instance was already passed on init
+        return self
+
+    def __enter__(self):
+        return self.driver
+
+    def __exit__(self, *args):
+        pass
+
+
 def save(
     self: GT,
     file: str,
     selector: str = "table",
     scale: float = 1.0,
     expand: int = 5,
-    web_driver: WebDrivers = "chrome",
+    web_driver: WebDrivers | webdriver.Remote = "chrome",
     window_size: tuple[int, int] = (6000, 6000),
     debug_port: None | int = None,
     encoding: str = "utf-8",
@@ -209,9 +227,12 @@ def save(
         (NOT IMPLEMENTED) The number of pixels to expand the screenshot by.  This can be
         increased to capture more of the surrounding area, or decreased to capture less.
     web_driver
-        The webdriver to use when taking the screenshot. By default, uses Google Chrome. Supports
-        `"firefox"` (Mozilla Firefox), `"safari"` (Apple Safari), and `"edge"` (Microsoft Edge).
-        Specified browser must be installed.
+        The webdriver to use when taking the screenshot. Either a driver name, or webdriver
+        instance. By default, uses Google Chrome. Supports `"firefox"` (Mozilla Firefox), `"safari"`
+        (Apple Safari), and `"edge"` (Microsoft Edge).
+
+        Specified browser must be installed. Note that if a webdriver instance is passed, options
+        that require setting up a webdriver, like debug_port, will not be used.
     window_size
         The size of the browser window to use when laying out the table. This shouldn't be necessary
         to capture a table, but may affect the tables appearance.
@@ -267,7 +288,11 @@ def save(
     html_content = as_raw_html(self)
 
     # Set the webdriver and options based on the chosen browser (`web_driver=` argument)
-    if web_driver == "chrome":
+    if isinstance(web_driver, webdriver.Remote):
+        wdriver = _NoOpDriverCtx(web_driver)
+        wd_options = None
+
+    elif web_driver == "chrome":
         wdriver = webdriver.Chrome
         wd_options = webdriver.ChromeOptions()
     elif web_driver == "safari":
@@ -279,6 +304,8 @@ def save(
     elif web_driver == "edge":
         wdriver = webdriver.Edge
         wd_options = webdriver.EdgeOptions()
+    else:
+        raise ValueError(f"Unsupported web driver: {web_driver}")
 
     # specify headless flag ----
     if web_driver in {"firefox", "edge"}:
