### 2024-10-02 b3191fc
feat: allow save file argument to be a Path
