### 2024-10-02 5af2702
Create test_utils_render_latex.py
