### 2024-10-02 a0b1b8f
Tweak logic of performing length conversions
