### 2024-10-02 b34f7ce
Add WidthDict class to store LaTeX colwidth data
