### 2024-10-02 c22a9b2
Modify names dealing with colwidth metadata
