### 2024-10-02 3dd29bf
Use WidthDict class within util fn
