### 2024-10-02 59c070f
Add untranslated code as commented code
