### 2024-10-02 1a81ae1
Add  several unit tests
