### 2024-10-02 d185604
Remove print statement
