### 2024-10-02 4296be9
Add logic to derive_table_width_statement_l()
