### 2024-10-03 a19dc8f
Update LENGTH_TRANSLATIONS_TO_PX dict with px val
