### 2024-10-03 ad0c45f
Add util functions for LaTeX rendering
