### 2024-10-03 941b72c
Add logic for `create_fontsize_statement_l()`
