### 2024-10-03 1bb1da5
Add logic for `create_columns_component_l()`
