+++ great_tables/_utils_render_latex.py
@@ -4,6 +4,8 @@ import re
 
 from ._gt_data import GTData
 from .quarto import check_quarto
+from great_tables._spanners import spanners_print_matrix
+from great_tables._utils_render_html import _get_spanners_matrix_height
 
 from typing import TypedDict, List
 
@@ -120,7 +122,7 @@ def create_width_dict_l(data: GTData) -> WidthDict:
     # Get the table width value
     tbl_width = data._options.table_width.value
 
-    # Get vector representation of stub layout
+    # Get list representation of stub layout
     stub_layout = data._stub._get_stub_layout(options=data._options)
 
     n = len(boxhead)
@@ -207,10 +209,210 @@ def create_width_dict_l(data: GTData) -> WidthDict:
     return width_dict
 
 
+def create_singlecolumn_width_text_l(pt: float, lw: float) -> str:
+
+    if pt <= 0 and lw <= 0:
+        out_txt = "0pt"
+    elif pt <= 0:
+        out_txt = "\\dimexpr {:.2f}\\linewidth -2\\tabcolsep-1.5\\arrayrulewidth".format(lw)
+    elif lw <= 0:
+        out_txt = "\\dimexpr {:.2f}pt -2\\tabcolsep-1.5\\arrayrulewidth".format(pt)
+    else:
+        out_txt = "\\dimexpr {:.2f}pt + {:.2f}\\linewidth -2\\tabcolsep-1.5\\arrayrulewidth".format(
+            pt, lw
+        )
+
+    return out_txt
+
+
+def calculate_multicolumn_width_text_l(
+    begins: list[str], ends: list[str], width_dict: WidthDict
+) -> str:
+
+    out_text = [""] * len(begins)
+
+    for i in range(len(begins)):
+
+        ind = range(begins[i], ends[i])
+
+        # If any of the combined columns has an unspecified width, it can't be determined
+        # so it is left blank
+        if any(width_dict["unspec"][i] > 0):
+
+            continue
+
+        pt_total = sum(width_dict["pt"][ind])
+        lw_total = sum(width_dict["lw"][ind])
+
+        if pt_total <= 0 & lw_total <= 0:
+
+            continue
+
+        else:
+
+            out_text[i] = create_singlecolumn_width_text_l(pt=pt_total, lw=lw_total)
+
+    return out_text
+
+
+def latex_heading_row(content: list[str]) -> str:
+
+    return "".join([" & ".join(content) + " \\\\ \n", "\\midrule\\addlinespace[2.5pt]\n"])
+
+
+def consolidate_cell_styles_l(styles_df: Any):
+    pass
+
+
 def create_table_start_l(data: GTData, width_dict: WidthDict) -> str:
 
-    # TODO: implement all logic
-    return ""
+    # Get list representation of stub layout
+    stub_layout = data._stub._get_stub_layout(options=data._options)
+
+    # Is the longtable environment being used?
+    latex_use_longtable = data._options.latex_use_longtable.value
+
+    # Extract only visible columns of `colwidth_df` based on stub_layout
+    types = ["default"]
+
+    if "rowname" in stub_layout:
+        types.append("stub")
+
+    if "group_label" in stub_layout:
+        types.append("row_group")
+
+    # Get indices of the types in `types` that are in the `type` key of `width_dict`
+    width_dict_visible_idx = [i for i, v in enumerate(width_dict["type"]) if v in types]
+
+    # Filter the `width_dict` dict entries based on the indices in `width_dict_visible_idx`
+    width_dict_visible = {k: [v[i] for i in width_dict_visible_idx] for k, v in width_dict.items()}
+
+    # Ensure that the `width_dict_visible` entries are sorted such that the
[…]
