### 2024-10-04 7f9ed37
fix: handle position or name to specify a group name location
