### 2024-10-04 f41f490
fix: regression when selecting all row groups
