### 2024-10-04 8c2b182
docs: restructure get-started into more sections
