### 2024-10-04 9fa9c56
tests: row and group locations via polars expression
