### 2024-10-04 977dcf9
fix: clean up and test column and spanner loc selection
