### 2024-10-04 25b6ad4
docs: add loc selection to get started
