+++ docs/get-started/loc-selection.qmd
@@ -1,0 +1,129 @@
+---
+title: Location selection
+jupyter: python3
+---
+
+```{python}
+# | echo: false
+import polars as pl
+from great_tables import GT
+
+data = [
+    ["header", "loc.header()", "composite"],
+    ["", "loc.title()", ""],
+    ["", "loc.subtitle()", ""],
+    ["boxhead", "loc.column_header()", "composite"],
+    ["", "loc.spanner_labels()", "columns"],
+    ["", "loc.column_labels()", "columns"],
+    ["row stub", "loc.stub()", "rows"],
+    ["", "loc.row_groups()", "rows"],
+    ["table body", "loc.body()", "columns and rows"],
+    ["footer", "loc.footer()", "composite"],
+    ["", "loc.source_notes()", ""],
+]
+
+df = pl.DataFrame(data, schema=["table part", "name", "selection"], orient="row")
+
+GT(df)
+```
+
+
+```{python}
+import polars as pl
+import polars.selectors as cs
+
+from great_tables import GT, loc, style, exibble
+
+pl_exibble = pl.from_pandas(exibble)[[0, 1, 4], ["num", "char", "group"]]
+```
+
+## simple locations
+
+```{python}
+(
+    GT(pl_exibble)
+    .tab_header("A title", "A subtitle")
+    .tab_style(
+        style.fill("yellow"),
+        loc.title(),
+    )
+)
+```
+
+## composite locations
+
+```{python}
+(
+    GT(pl_exibble)
+    .tab_header("A title", "A subtitle")
+    .tab_style(
+        style.fill("yellow"),
+        loc.header(),
+    )
+)
+```
+
+## body columns and rows
+
+```{python}
+(
+    GT(pl_exibble).tab_style(
+        style.fill("yellow"),
+        loc.body(
+            columns=cs.starts_with("cha"),
+            rows=pl.col("char").str.contains("a"),
+        ),
+    )
+)
+```
+
+## column labels
+
+```{python}
+GT(pl_exibble).tab_style(
+    style.fill("yellow"),
+    loc.column_labels(
+        cs.starts_with("cha"),
+    ),
+)
+```
+
+## row and group names
+
+* by name
+* by index
+* by expression
+
+```{python}
+gt = GT(pl_exibble).tab_stub(
+    rowname_col="char",
+    groupname_col="group",
+)
+
+gt.tab_style(style.fill("yellow"), loc.stub())
+```
+
+
+```{python}
+gt.tab_style(style.fill("yellow"), loc.stub("banana"))
+```
+
+```{python}
+gt.tab_style(style.fill("yellow"), loc.stub(["apricot", 2]))
+```
+
+### groups by name and position
+
+```{python}
+gt.tab_style(
+    style.fill("yellow"),
+    loc.row_groups("grp_b"),
+)
+```
+
+```{python}
+gt.tab_style(
+    style.fill("yellow"),
+    loc.row_groups(1),
+)
+```
