### 2024-10-04 fb5709f
docs: add more narrative
