### 2024-10-05 62042f4
Add `GT.write_html()` function
