### 2024-10-05 135b232
Add tests for `GT.write_html()` function
