### 2024-10-05 ecff5da
update _quarto.yml to expose `GT.write_html()`
