### 2024-10-05 e627865
Add PyArrow.Table support
