+++ great_tables/_tbl_data.py
@@ -54,6 +54,9 @@ else:
     class PlDataFrame(AbstractBackend):
         _backends = [("polars", "DataFrame")]
 
+    class PyArrowTable(AbstractBackend):
+        _backends = [("pyarrow", "Table")]
+
     class PlSelectExpr(AbstractBackend):
         _backends = [("polars.selectors", "_selector_proxy_")]
 
@@ -65,6 +68,9 @@ else:
 
     class PlSeries(AbstractBackend):
         _backends = [("polars", "Series")]
+
+    class PyArrowArray(AbstractBackend):
+        _backends = [("pyarrow", "Array")]
 
     class PdNA(AbstractBackend):
         _backends = [("pandas", "NA")]
@@ -85,8 +91,10 @@ else:
 
     DataFrameLike.register(PdDataFrame)
     DataFrameLike.register(PlDataFrame)
+    DataFrameLike.register(PyArrowTable)
     SeriesLike.register(PdSeries)
     SeriesLike.register(PlSeries)
+    SeriesLike.register(PyArrowArray)
 
     TblData = DataFrameLike
 
@@ -140,6 +148,12 @@ def _(data: PdDataFrame):
 def _(data: PlDataFrame):
     return data.clone()
 
+@copy_data.register(PyArrowTable)
+def _(data: PyArrowTable):
+    import pyarrow as pa
+
+    return pa.table(data)
+
 
 # get_column_names ----
 @singledispatch
@@ -157,6 +171,9 @@ def _(data: PdDataFrame):
 def _(data: PlDataFrame):
     return data.columns
 
+@get_column_names.register(PyArrowTable)
+def _(data: PyArrowTable):
+    return data.column_names
 
 # n_rows ----
 
@@ -172,6 +189,9 @@ def n_rows(data: DataFrameLike) -> int:
 def _(data: Any) -> int:
     return len(data)
 
+@n_rows.register(PyArrowTable)
+def _(data: PyArrowTable) -> int:
+    return data.num_rows
 
 # _get_cell ----
 
@@ -197,6 +217,9 @@ def _(data: Any, row: int, col: str) -> Any:
 
     return data.iloc[row, col_ii]
 
+@_get_cell.register(PyArrowTable)
+def _(data: PyArrowTable, row: int, column: str) -> Any:
+    return data.column(column).take([row]).to_pylist()[0]
 
 # _set_cell ----
 
@@ -218,15 +241,22 @@ def _(data, row: int, column: str, value: Any) -> None:
 def _(data, row: int, column: str, value: Any) -> None:
     data[row, column] = value
 
+@_set_cell.register(PyArrowTable)
+def _(data: PyArrowTable, row: int, column: str, value: Any) -> None:
+    raise NotImplementedError("Setting values in PyArrow tables is not supported.")
 
 # _get_column_dtype ----
 
 
 @singledispatch
-def _get_column_dtype(data: DataFrameLike, column: str) -> str:
+def _get_column_dtype(data: DataFrameLike, column: str) -> Any:
     """Get the data type for a single column in the input data table"""
     return data[column].dtype
 
+
+@_get_column_dtype.register(PyArrowTable)
+def _(data: PyArrowTable, column: str) -> Any:
+    return data.column(column).type
 
 # reorder ----
 
@@ -248,6 +278,10 @@ def _(data: PdDataFrame, rows: list[int], columns: list[str]) -> PdDataFrame:
 @reorder.register
 def _(data: PlDataFrame, rows: list[int], columns: list[str]) -> PlDataFrame:
     return data[rows, columns]
+
+@reorder.register
+def _(data: PyArrowTable, rows: list[int], columns: list[str]) -> PyArrowTable:
[…]
