### 2024-10-05 6477bcb
Update substitutions too
