+++ great_tables/_gt_data.py
@@ -174,7 +174,7 @@ class Body:
                 # TODO: I think that this is very inefficient with polars, so
                 # we could either accumulate results and set them per column, or
                 # could always use a pandas DataFrame inside Body?
-                _set_cell(self.body, row, col, result)
+                self.body = _set_cell(self.body, row, col, result)
 
         return self
 
