### 2024-10-05 ed8e8e0
Update function signature
