### 2024-10-05 e501880
Few typing fixes
