### 2024-10-07 025ac69
Remove usage of the `os` module in tests for `GT.write_html()`
