+++ docs/blog/introduction-0.13.0/index.qmd
@@ -1,0 +1,226 @@
+---
+title: "Great Tables `v0.13.0`: Styles all over the table"
+html-table-processing: none
+author: Rich Iannone and Michael Chow
+date: 2024-10-08
+freeze: true
+jupyter: python3
+---
+
+We did something in Great Tables (`0.13.0`) that'll make your tables that much more customizable: *super* fine-grained ways of setting styles throughout the table. Before you were largely constrained to styling through the following strategies:
+
+1. use a limited set of styles (e.g., background color, font weight, etc.) to different table locations like the stub, the column labels, etc., through `tab_options()`
+2. use `tab_style()` with a larger set of styling options for the table body cells (specified by `loc.body()`)
+
+In `v0.13.0`, we can target much more than just the table body! Here is the expanded set of `loc.*()` methods along with the locations that they can target.
+
+- `loc.title()`: the table title
+- `loc.subtitle()`: the table subtitle
+- `loc.stubhead()`: the stubhead
+- `loc.spanner_labels()`: the spanner labels
+- `loc.column_labels()`: the column labels
+- `loc.stub()`: the table stub
+- `loc.row_groups()`: the row group labels
+- `loc.body()`: data cells in the table body
+- `loc.source_notes()`: the source notes
+
+This hyperexpansion of the `loc` module to all locations in the table means that there won't be a spot in the table you can't add styling to. And this is great because it gives you free reign fully customize the look of your table.
+
+Let's make a table and see how this new feature could be used.
+
+### Starting things off with a big GT table
+
+The table we'll make uses the `nuclides` dataset (available in the `great_tables.data` module). Through use of the `tab_*()` methods, quite a few table components (hence *locations*) will be added.
+
+```{python}
+
+# | code-fold: true
+# | code-summary: "Show the code"
+
+from great_tables import GT, md, style, loc, google_font
+from great_tables.data import nuclides
+import polars as pl
+import polars.selectors as cs
+
+nuclides_mini = (
+    pl.from_pandas(nuclides)
+    .filter(pl.col("element") == "C")
+    .with_columns(pl.col("nuclide").str.replace(r"[0-9]+$", ""))
+    .with_columns(mass_number=pl.col("z") + pl.col("n"))
+    .with_columns(
+        isotope=pl.concat_str(pl.col("element") + "-" + pl.col("mass_number").cast(pl.String))
+    )
+    .select(["isotope", "atomic_mass", "half_life", "isospin", "decay_1", "decay_2", "decay_3"])
+)
+
+gt_tbl = (
+    GT(nuclides_mini, rowname_col="isotope")
+    .tab_header(
+        title="Isotopes of Carbon",
+        subtitle="There are two stable isotopes of carbon and twelve that are unstable.",
+    )
+    .tab_spanner(label="Decay Mode", columns=cs.starts_with("decay"))
+    .tab_source_note(md("Data obtained from the *nuclides* dataset."))
+    .tab_stubhead(label="Isotope")
+    .fmt_scientific(columns="half_life")
+    .fmt_number(
+        columns="atomic_mass",
+        decimals=4,
+        scale_by=1 / 1e6,
+    )
+    .sub_missing(columns="half_life", missing_text=md("**STABLE**"))
+    .sub_missing(columns=cs.starts_with("decay"))
+    .cols_label(
+        atomic_mass="Atomic Mass",
+        half_life="Half Life, s",
+        decay_1="1",
+        decay_2="2",
+        decay_3="3",
+    )
[…]
