+++ docs/blog/introduction-0.13.0/index.qmd
@@ -1,5 +1,5 @@
 ---
-title: "Great Tables `v0.13.0`: Styles all over the table"
+title: "Great Tables `v0.13.0`: Applying styles to all table locations"
 html-table-processing: none
 author: Rich Iannone and Michael Chow
 date: 2024-10-08
