### 2024-10-08 7707bb1
Update docs/blog/introduction-0.13.0/index.qmd
Co-authored-by: Michael Chow <mc_al_github@fastmail.com>
