### 2024-10-09 8c27676
Remove list of locations
