### 2024-10-09 d40040b
Consolidate set_cell immutability behavior
