### 2024-10-09 d52bbf5
Make some text adjustments
