### 2024-10-10 4f2ae11
dev: update to quartodoc v0.8.0
