+++ .github/workflows/ci-docs.yaml
@@ -19,7 +19,6 @@ jobs:
           python-version: "3.10"
       - name: Install dependencies
         run: |
-          python -m pip install git+https://github.com/machow/quartodoc.git
           python -m pip install ".[all]"
       - uses: quarto-dev/quarto-actions/setup@v2
       - name: Build docs
