### 2024-10-11 3d2e580
docs: update quartodoc to v0.8.1
