+++ pyproject.toml
@@ -66,7 +66,7 @@ dev = [
 dev-no-pandas = [
     "black",
     "jupyter",
-    "quartodoc>=0.8.0; python_version >= '3.9'",
+    "quartodoc>=0.8.1; python_version >= '3.9'",
     "griffe==0.38.1",
     "polars",
     "pre-commit==2.15.0",
