### 2024-10-15 ceb8e60
Leverage the parameters of `GT.as_raw_html()`
