### 2024-10-15 d8d5e8f
Add to imports for LaTeX rendering
