### 2024-10-15 856e636
Remove unneeded code
