### 2024-10-15 69e7d6e
Remove terminating newline from heading row stmt
