### 2024-10-15 831299f
Make correction to dict comprehension
