### 2024-10-15 4f25112
Remove terminating newline from table start stmt
