### 2024-10-15 63fdd63
Use pass instead of returning "" for a caption
