### 2024-10-15 5b27b35
Add logic for creating a LaTeX table heading
