### 2024-10-15 7a1040c
Add logic for creating a LaTeX table body
