### 2024-10-15 acf6e87
Add logic for creating a LaTeX table footer
