### 2024-10-15 43345cf
Remove unnecessary newlines
