### 2024-10-15 07ce01f
Remove unnecessary newlines in tbl start stmt
