### 2024-10-15 26e477f
Remove all newlines in wrap_end stmt
