### 2024-10-15 e7f07c3
Raise NotImplemented if using a stub or row groups
