### 2024-10-15 cc8c858
Remove stmt that creates a caption component
