### 2024-10-15 08b6c78
Obtain the 'use_longtable' option
