### 2024-10-15 f5875e4
Remove redundant variable from 'finalized_table'
