### 2024-10-15 269825c
Remove 'caption_component' from built LaTeX table
