### 2024-10-15 350cb9d
Make corrections to two LaTeX test cases
