### 2024-10-16 ad35bdf
Add the `escape_latex()` util function
