### 2024-10-16 ef88710
Add basic tests for `escape_latex()`
