+++ tests/test_utils_render_latex.py
@@ -16,6 +16,7 @@ from great_tables._utils_render_latex import (
     create_width_dict_l,
     create_wrap_start_l,
     create_fontsize_statement_l,
+    escape_latex,
 )
 
 
@@ -121,6 +122,13 @@ def test_create_width_dict_l_settings():
         "left",
     ]
     assert width_dict["tbl_width"] is None
+
+
+def test_escape_latex():
+
+    assert escape_latex("a & b") == "a \\& b"
+    assert escape_latex("a & b & c") == "a \\& b \\& c"
+    assert escape_latex("\\a_\\d") == "\\\\a\\_\\\\d"
 
 
 def test_create_fontsize_statement_l():
