### 2024-10-16 927e4ac
Add the `_context_minus_mark()` util fn
