### 2024-10-16 dd92e10
Add the `_context_percent_mark()` util fn
