+++ great_tables/_formats.py
@@ -2695,6 +2695,15 @@ def _context_minus_mark(context: str) -> str:
     return mark
 
 
+def _context_percent_mark(context: str) -> str:
+
+    if context == "latex":
+        mark = "\\%"
+    else:
+        mark = "%"
+
+    return mark
+
 
 def _replace_minus(string: str, minus_mark: str) -> str:
     """
