### 2024-10-16 09215b6
Add the `_context_dollar_mark()` util fn
