### 2024-10-16 4738acb
Remove temporary version of `_context_minus_mark()`
