### 2024-10-16 fd228a6
Create temp copies of fmt fns for html/latex contexts
