+++ great_tables/_formats.py
@@ -335,7 +335,7 @@ def fmt_number(
 
         # Implement minus sign replacement for `x_formatted`
         if is_negative:
-            minus_mark = _context_minus_mark()
+            minus_mark = _context_minus_mark(context="html")
             x_formatted = _replace_minus(x_formatted, minus_mark=minus_mark)
 
         # Use a supplied pattern specification to decorate the formatted value
@@ -344,7 +344,71 @@ def fmt_number(
 
         return x_formatted
 
-    return fmt(self, fns=fmt_number_fn, columns=columns, rows=rows)
+    # Generate a function that will operate on single `x` values in the table body
+    def fmt_number_fn_latex(
+        x: float | None,
+        decimals: int = decimals,
+        n_sigfig: int | None = n_sigfig,
+        drop_trailing_zeros: bool = drop_trailing_zeros,
+        drop_trailing_dec_mark: bool = drop_trailing_dec_mark,
+        use_seps: bool = use_seps,
+        scale_by: float = scale_by,
+        compact: bool = compact,
+        sep_mark: str = sep_mark,
+        dec_mark: str = dec_mark,
+        force_sign: bool = force_sign,
+    ):
+        if is_na(self._tbl_data, x):
+            return x
+
+        # Scale `x` value by a defined `scale_by` value
+        x = x * scale_by
+
+        # Determine whether the value is positive
+        is_negative = _has_negative_value(value=x)
+
+        if compact:
+            x_formatted = _format_number_compactly(
+                value=x,
+                decimals=decimals,
+                n_sigfig=n_sigfig,
+                drop_trailing_zeros=drop_trailing_zeros,
+                drop_trailing_dec_mark=drop_trailing_dec_mark,
+                use_seps=use_seps,
+                sep_mark=sep_mark,
+                dec_mark=dec_mark,
+                force_sign=force_sign,
+            )
+        else:
+            x_formatted = _value_to_decimal_notation(
+                value=x,
+                decimals=decimals,
+                n_sigfig=n_sigfig,
+                drop_trailing_zeros=drop_trailing_zeros,
+                drop_trailing_dec_mark=drop_trailing_dec_mark,
+                use_seps=use_seps,
+                sep_mark=sep_mark,
+                dec_mark=dec_mark,
+                force_sign=force_sign,
+            )
+
+        # Implement minus sign replacement for `x_formatted`
+        if is_negative:
+            minus_mark = _context_minus_mark(context="latex")
+            x_formatted = _replace_minus(x_formatted, minus_mark=minus_mark)
+
+        # Use a supplied pattern specification to decorate the formatted value
+        if pattern != "{x}":
+            x_formatted = pattern.replace("{x}", x_formatted)
+
+        return x_formatted
+
+    return fmt(
+        self,
+        fns=FormatFns(html=fmt_number_fn, latex=fmt_number_fn_latex, default=fmt_number_fn),
+        columns=columns,
+        rows=rows,
+    )
 
 
 def fmt_integer(
@@ -501,7 +565,7 @@ def fmt_integer(
 
         # Implement minus sign replacement for `x_formatted`
         if is_negative:
-            minus_mark = _context_minus_mark()
+            minus_mark = _context_minus_mark(context="html")
             x_formatted = _replace_minus(x_formatted, minus_mark=minus_mark)
 
         # Use a supplied pattern specification to decorate the formatted value
@@ -687,7 +751,7 @@ def fmt_scientific(
         # Determine whether the value is positive
         is_positive = _has_positive_value(value=x)
 
-        minus_mark = _context_minus_mark()
+        minus_mark = _context_minus_mark(context="html")
 
         x_sci_notn = _value_to_scientific_notation(
             value=x,
[…]
