### 2024-10-16 6270957
Apply `escape_latex()` to column labels
