### 2024-10-16 b5fb89b
Define more contexts in `_context_exp_marks()`
