+++ great_tables/_formats.py
@@ -2879,8 +2879,16 @@ def _has_sci_order_zero(value: int | float) -> bool:
     return (value >= 1 and value < 10) or (value <= -1 and value > -10) or value == 0
 
 
-def _context_exp_marks() -> list[str]:
-    return [" \u00D7 10<sup style='font-size: 65%;'>", "</sup>"]
+def _context_exp_marks(context: str) -> list[str]:
+
+    if context == "html":
+        marks = [" \u00D7 10<sup style='font-size: 65%;'>", "</sup>"]
+    elif context == "latex":
+        marks = [" $\\times$ 10\\textsuperscript{", "}"]
+    else:
+        marks = [" \u00D7 10^", ""]
+
+    return marks
 
 
 def _context_exp_str(exp_style: str) -> str:
