### 2024-10-16 65a2826
Use context-aware functions for many fmt methods
