### 2024-10-16 823fdfc
Redefine function signature
