+++ great_tables/_utils_render_latex.py
@@ -229,7 +229,7 @@ def latex_heading_row(content: list[str]) -> str:
     return "".join([" & ".join(content) + " \\\\ \n", "\\midrule\\addlinespace[2.5pt]"])
 
 
-def consolidate_cell_styles_l(styles_df: Any):
+def consolidate_cell_styles_l():
     pass
 
 
