### 2024-10-16 695490d
Remove unneeded variable
