### 2024-10-16 27c1a32
Remove unused code from LaTeX column building
