+++ great_tables/_utils_render_latex.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import re
 
-from ._gt_data import GTData
+from ._gt_data import GTData, GroupRowInfo
 from ._tbl_data import _get_cell, cast_frame_to_string, replace_null_frame
 from .quarto import check_quarto
 from great_tables._spanners import spanners_print_matrix
