### 2024-10-16 f6800b9
Remove type of Any for `cell_content`
