### 2024-10-16 82328f6
Add LaTeX examples page to docs site
