### 2024-10-16 ff8051c
Add LaTeX examples page to top navbar
