+++ docs/_quarto.yml
@@ -17,6 +17,8 @@ website:
         text: Reference
       - href: blog/index.qmd
         text: Blog
+      - text: LaTeX Examples
+        file: latex_examples/index.qmd
     right:
       - icon: github
         href: https://github.com/posit-dev/great-tables
