### 2024-10-16 82433a2
Move LaTeX examples to lower alphanumeric dir name
