### 2024-10-16 f34f2b1
Ensure that tinytex is installed after quarto install
