+++ .github/workflows/ci-docs.yaml
@@ -21,6 +21,8 @@ jobs:
         run: |
           python -m pip install ".[all]"
       - uses: quarto-dev/quarto-actions/setup@v2
+        with:
+          tinytex: true
       - name: Build docs
         run: |
           make docs-build
