### 2024-10-16 0dd0109
Pull table_width from width_dict
