### 2024-10-16 7606069
Update LaTeX examples .qmd file
