### 2024-10-16 30625c4
Update LaTeX examples .qmd file
