### 2024-10-17 0061441
Enlarge scope of `process_text()` to multiple contexts
