+++ great_tables/_text.py
@@ -26,24 +26,54 @@ def _md_html(x: str) -> str:
     return re.sub(r"^<p>|</p>\n$", "", str)
 
 
-def _process_text(x: str | Text | None) -> str:
+def _md_latex(x: str) -> str:
+
+    # TODO: Implement commonmark to LaTeX conversion (through a different library as
+    # commonmark-py does not support it)
+
+    return x
+
+
+def _process_text(x: str | Text | None, context: str = "html") -> str:
+
     from great_tables._helpers import UnitStr
 
     if x is None:
         return ""
 
-    if isinstance(x, Md):
-        return _md_html(x.text)
-    elif isinstance(x, Html):
-        return x.text
-    elif isinstance(x, str):
-        return _html_escape(x)
-    elif isinstance(x, Text):
-        return x.text
-    elif isinstance(x, UnitStr):
-        return x.to_html()
+    if context == "html":
+
+        if isinstance(x, Md):
+            return _md_html(x.text)
+        elif isinstance(x, Html):
+            return x.text
+        elif isinstance(x, str):
+            return _html_escape(x)
+        elif isinstance(x, Text):
+            return x.text
+        elif isinstance(x, UnitStr):
+            return x.to_html()
+        else:
+            raise TypeError(f"Invalid type: {type(x)}")
+
+    elif context == "latex":
+
+        if isinstance(x, Md):
+            return _md_latex(x.text)
+        elif isinstance(x, Html):
+            return x.text
+        elif isinstance(x, str):
+            return x
+        elif isinstance(x, Text):
+            return x.text
+        elif isinstance(x, UnitStr):
+            # TODO: this is currently not implemented
+            return x.to_latex()
+        else:
+            raise TypeError(f"Invalid type: {type(x)}")
+
     else:
-        raise TypeError(f"Invalid type: {type(x)}")
+        raise ValueError(f"Invalid context: {context}")
 
 
 def _process_text_id(x: str | Text | None) -> str:
