### 2024-10-17 874e792
Implement spanners logic in LaTeX tables
