### 2024-10-17 f9f1a8a
Update LaTeX examples .qmd file
