+++ docs/a-latex_examples/index.qmd
@@ -29,6 +29,14 @@ print(
             # groupname_col="mfr",
         )
         .tab_header(title="The title", subtitle="The subtitle")
+        .tab_spanner(
+            label="Make and Model",
+            columns=["mfr", "model"]
+        )
+        .tab_spanner(
+            label="Performance Characteristics",
+            columns=["hp", "trq"]
+        )
         .tab_source_note(source_note="A table note.")
         .tab_source_note(source_note="A second table note.")
         .fmt_scientific(columns="trq", decimals=3)
