### 2024-10-17 6835cd5
Move away from f-string when not necessary
