### 2024-10-17 74ea075
Simplify example with spanner label
