### 2024-10-17 38e1286
Increase test coverage and fix group_splits
