### 2024-10-17 1caf472
Make corrections to spanner formatting in LaTeX
