### 2024-10-17 703fa37
Update LaTeX examples in .qmd file
