### 2024-10-17 ad0c4f5
Add tests for creating LaTeX table heading
