### 2024-10-17 98fed5c
Raise NotImplementedError for md-to-latex conversion
