### 2024-10-17 7697bd3
Raise NotImplementedError for using units w/ LaTeX
