+++ great_tables/_helpers.py
@@ -851,6 +851,9 @@ class UnitStr:
         return built_units
 
     # TODO: add `to_latex` method
+    def to_latex(self) -> str:
+
+        raise NotImplementedError("LaTeX conversion of units is not yet supported.")
 
     def _repr_html_(self):
         return self.to_html()
