### 2024-10-17 934a6dc
Remove certain TODOs and commented code
