### 2024-10-17 e99b512
Add several tests for spanners in LaTeX outputs
