### 2024-10-17 5913f3b
Add complicated test of spanners in LaTeX render
