### 2024-10-17 53cf254
Add several tests for formatting in LaTeX
