### 2024-10-18 3b86a57
Add several tests for formatting in LaTeX
