### 2024-10-18 7fb7fb9
Add fn to selectively escape for LaTeX (for pattern str)
