### 2024-10-18 5809987
Add tests for `escape_pattern_str_latex()`
