+++ tests/test_utils_render_latex.py
@@ -21,6 +21,7 @@ from great_tables._utils_render_latex import (
     create_body_component_l,
     create_columns_component_l,
     create_wrap_end_l,
+    escape_pattern_str_latex,
 )
 
 
@@ -148,6 +149,12 @@ def test_escape_latex():
     assert escape_latex("a & b") == "a \\& b"
     assert escape_latex("a & b & c") == "a \\& b \\& c"
     assert escape_latex("\\a_\\d") == "\\\\a\\_\\\\d"
+
+
+def test_escape_pattern_str_latex():
+
+    assert escape_pattern_str_latex("{x}") == "{x}"
+    assert escape_pattern_str_latex("a $_{1} %ab {2}") == "a \\$\\_{1} \\%ab {2}"
 
 
 def test_create_fontsize_statement_l():
