### 2024-10-18 b162aee
Add test for `escape_pattern_str_latex()`
