### 2024-10-18 d936f5e
Perform LaTeX-escaping of pattern in several fmt methods
