+++ great_tables/_formats.py
@@ -29,6 +29,7 @@ from ._tbl_data import (
 )
 from ._text import _md_html
 from ._utils import _str_detect, _str_replace
+from ._utils_render_latex import escape_pattern_str_latex
 from ._utils_nanoplots import _generate_nanoplot
 
 
@@ -401,6 +402,11 @@ def fmt_number_context(
 
     # Use a supplied pattern specification to decorate the formatted value
     if pattern != "{x}":
+
+        # Escape LaTeX special characters from literals in the pattern
+        if context == "latex":
+            pattern = escape_pattern_str_latex(pattern_str=pattern)
+
         x_formatted = pattern.replace("{x}", x_formatted)
 
     return x_formatted
@@ -613,6 +619,11 @@ def fmt_integer_context(
 
     # Use a supplied pattern specification to decorate the formatted value
     if pattern != "{x}":
+
+        # Escape LaTeX special characters from literals in the pattern
+        if context == "latex":
+            pattern = escape_pattern_str_latex(pattern_str=pattern)
+
         x_formatted = pattern.replace("{x}", x_formatted)
 
     return x_formatted
@@ -929,6 +940,11 @@ def fmt_scientific_context(
 
     # Use a supplied pattern specification to decorate the formatted value
     if pattern != "{x}":
+
+        # Escape LaTeX special characters from literals in the pattern
+        if context == "latex":
+            pattern = escape_pattern_str_latex(pattern_str=pattern)
+
         x_formatted = pattern.replace("{x}", x_formatted)
 
     return x_formatted
@@ -1207,6 +1223,11 @@ def fmt_percent_context(
 
     # Use a supplied pattern specification to decorate the formatted value
     if pattern != "{x}":
+
+        # Escape LaTeX special characters from literals in the pattern
+        if context == "latex":
+            pattern = escape_pattern_str_latex(pattern_str=pattern)
+
         x_formatted = pattern.replace("{x}", x_formatted)
 
     return x_formatted
@@ -1507,6 +1528,11 @@ def fmt_currency_context(
 
     # Use a supplied pattern specification to decorate the formatted value
     if pattern != "{x}":
+
+        # Escape LaTeX special characters from literals in the pattern
+        if context == "latex":
+            pattern = escape_pattern_str_latex(pattern_str=pattern)
+
         x_formatted = pattern.replace("{x}", x_formatted)
 
     return x_formatted
