### 2024-10-18 07b39d4
Use context-aware fns for more fmt methods
