+++ great_tables/_formats.py
@@ -1455,7 +1455,6 @@ def fmt_currency(
     )
 
 
-# Generate a function that will operate on single `x` values in the table body
 def fmt_currency_context(
     x: float | None,
     data: GTData,
@@ -1602,50 +1601,85 @@ def fmt_roman(
     # Check that the `case` value is valid and only consists of the string 'upper' or 'lower'
     _validate_case(case=case)
 
-    # Generate a function that will operate on single `x` values in the table body
-    def fmt_roman_fn(
-        x: float,
-        case: str = case,
-    ):
-        # If the `x` value is a Pandas 'NA', then return the same value
-        if is_na(self._tbl_data, x):
-            return x
+    # TODO: fix type errors for members of `FormatFns`
+    return fmt(
+        self,
+        fns=FormatFns(
+            html=partial(
+                fmt_roman_context,
+                data=self,
+                case=case,
+                pattern=pattern,
+                context="html",
+            ),
+            latex=partial(
+                fmt_roman_context,
+                data=self,
+                case=case,
+                pattern=pattern,
+                context="latex",
+            ),
+            default=partial(
+                fmt_roman_context,
+                data=self,
+                case=case,
+                pattern=pattern,
+                context="html",
+            ),
+        ),
+        columns=columns,
+        rows=rows,
+    )
 
-        # Get the absolute value of `x` so that negative values are handled
-        x = abs(x)
 
-        # Round x to 0 digits with the R-H-U method of rounding (for reproducibility purposes)
-        x = _round_rhu(x, 0)
+def fmt_roman_context(
+    x: float,
+    data: GTData,
+    case: str,
+    pattern: str,
+    context: str,
+):
+    if is_na(data._tbl_data, x):
+        return x
 
-        # Determine if `x` is in the range of 1 to 3899 and if it is zero
-        x_is_in_range = x > 0 and x < 3900
-        x_is_zero = x == 0
+    # Get the absolute value of `x` so that negative values are handled
+    x = abs(x)
 
-        if not x_is_in_range and not x_is_zero:
-            # We cannot format a 'large' integer to roman numerals, so we return a string
-            # that indicates this
-            return "ex terminis"
-        elif x_is_zero:
-            # Zero is a special case and is handled separately with the character 'N'
-            # which stands for 'nulla' (i.e., 'nothing')
-            x_formatted = "N"
-        else:
-            # All other values are formatted with the `_as_roman()` utility function
-            x_formatted = _as_roman(x)
+    # Round x to 0 digits with the R-H-U method of rounding (for reproducibility purposes)
+    x = _round_rhu(x, 0)
 
-        # Transform the case of the formatted value
-        if case == "upper":
-            pass
-        else:
-            x_formatted = x_formatted.lower()
+    # Determine if `x` is in the range of 1 to 3899 and if it is zero
+    x_is_in_range = x > 0 and x < 3900
+    x_is_zero = x == 0
 
-        # Use a supplied pattern specification to decorate the formatted value
-        if pattern != "{x}":
-            x_formatted = pattern.replace("{x}", x_formatted)
+    if not x_is_in_range and not x_is_zero:
+        # We cannot format a 'large' integer to roman numerals, so we return a string
+        # that indicates this
+        return "ex terminis"
+    elif x_is_zero:
+        # Zero is a special case and is handled separately with the character 'N'
+        # which stands for 'nulla' (i.e., 'nothing')
+        x_formatted = "N"
+    else:
+        # All other values are formatted with the `_as_roman()` utility function
+        x_formatted = _as_roman(x)
 
-        return x_formatted
+    # Transform the case of the formatted value
+    if case == "upper":
[…]
