### 2024-10-18 a8c0c8d
Add example to LaTeX .qmd file
