### 2024-10-18 0785401
Make correction to LaTeX examples .qmd file
