+++ docs/a-latex_examples/index.qmd
@@ -8,7 +8,7 @@ format: pdf
 
 import polars as pl
 from great_tables import GT
-from great_tables.data import gtcars, islands, airquality
+from great_tables.data import gtcars, islands, airquality, constants
 ```
 
 ```{python}
