### 2024-10-18 a70890e
Move `process_string()` fn to `_utils.py`
