### 2024-10-18 97a7096
Add docstring for the `process_string()` util fn
