### 2024-10-18 cbf0929
Add docstrings for several LaTeX-rendering fns
