+++ great_tables/_utils_render_latex.py
@@ -241,6 +241,25 @@ def consolidate_cell_styles_l():
 
 
 def create_table_start_l(data: GTData, width_dict: WidthDict) -> str:
+    """
+    Create the table start component for LaTeX output.
+
+    This function generates the LaTeX code that signifies the start of the table. The output is
+    different depending on whether the table uses the `longtable` environment or not.
+
+    Parameters
+    ----------
+    data : GTData
+        The GTData object that contains all the information about the table.
+    width_dict : WidthDict
+        A dictionary that contains information about the widths of the columns and the table itself.
+        This dictionary is generated by the `create_width_dict_l()` function.
+
+    Returns
+    -------
+    str
+        The LaTeX code that signifies the start of the table.
+    """
 
     # Get list representation of stub layout
     stub_layout = data._stub._get_stub_layout(options=data._options)
@@ -406,6 +425,23 @@ def create_caption_component_l(data: GTData) -> str:
 
 
 def create_heading_component_l(data: GTData) -> str:
+    """
+    Create the heading component for LaTeX output.
+
+    This function generates the LaTeX code for the heading component of a table which involves the
+    title and the optional subtitle. There is variation in the output based on whether the table
+    uses the `longtable` environment or not.
+
+    Parameters
+    ----------
+    data : GTData
+        The GTData object that contains all the information about the table.
+
+    Returns
+    -------
+    str
+        The LaTeX code for the heading component of the table.
+    """
 
     title = data._heading.title
     subtitle = data._heading.subtitle
@@ -444,6 +480,25 @@ def create_heading_component_l(data: GTData) -> str:
 
 
 def create_columns_component_l(data: GTData, width_dict: WidthDict) -> str:
+    """
+    Create the columns component for LaTeX output.
+
+    This function generates the LaTeX code for the columns component of a table which involves the
+    column headings and the spanners.
+
+    Parameters
+    ----------
+    data : GTData
+        The GTData object that contains all the information about the table.
+    width_dict : WidthDict
+        A dictionary that contains information about the widths of the columns and the table itself.
+        This dictionary is generated by the `create_width_dict_l()` function.
+
+    Returns
+    -------
+    str
+        The LaTeX code for the columns component of the table.
+    """
 
     # Get list representation of stub layout
     stub_layout = data._stub._get_stub_layout(options=data._options)
@@ -640,6 +695,25 @@ def create_columns_component_l(data: GTData, width_dict: WidthDict) -> str:
 
 
 def create_body_component_l(data: GTData, width_dict: WidthDict) -> str:
+    """
+    Create the body component for LaTeX output.
+
+    This function generates the LaTeX code for the body component of a table which involves the
+    data cells, the row groups, and the stub.
+
+    Parameters
+    ----------
+    data : GTData
+        The GTData object that contains all the information about the table.
+    width_dict : WidthDict
+        A dictionary that contains information about the widths of the columns and the table itself.
+        This dictionary is generated by the `create_width_dict_l()` function.
+
+    Returns
+    -------
+    str
+        The LaTeX code for the body component of the table.
+    """
 
     _str_orig_data = cast_frame_to_string(data._tbl_data)
     tbl_data = replace_null_frame(data._body.body, _str_orig_data)
@@ -697,6 +771,22 @@ def create_body_component_l(data: GTData, width_dict: WidthDict) -> str:
 
 
 def create_footer_component_l(data: GTData) -> str:
+    """
+    Create the footer component for LaTeX output.
+
+    This function generates the LaTeX code for the footer component of a table which involves the
+    source notes.
+
+    Parameters
+    ----------
+    data : GTData
+        The GTData object that contains all the information about the table.
+
+    Returns
+    -------
[…]
