### 2024-10-18 9235d97
Add tests for creation of LaTeX table footer
