### 2024-10-18 2f132c8
Add DataFrame fixture for datetime formatting
