### 2024-10-18 baccce1
Add tests for date and time formatting in LaTeX
