### 2024-10-18 4a21981
Add test for datetime formatting in LaTeX
