### 2024-10-19 3d9b486
Add the `_latex_escape()` util fn
