+++ great_tables/_text.py
@@ -81,3 +81,11 @@ def _process_text_id(x: str | Text | None) -> str:
 
 def _html_escape(x: str) -> str:
     return html.escape(x)
+
+
+def _latex_escape(text: str) -> str:
+
+    latex_escape_regex = "[\\\\&%$#_{}~^]"
+    text = re.sub(latex_escape_regex, lambda match: "\\" + match.group(), text)
+
+    return text
