### 2024-10-19 ffc1889
Use `_latex_escape()` within `_process_text()`
