### 2024-10-19 c5ccc34
Add the `_get_visible_cells()` util fn
