### 2024-10-19 a545adb
Add `migrate_unformatted_to_output()` util fn
