### 2024-10-19 76c0298
When building LaTeX table, process unformatted text
