### 2024-10-19 61719be
Use `_process_text()` when rendering LaTeX
