### 2024-10-19 363e13c
Create test_text.py
