### 2024-10-19 a2047a6
Add `save_pdf` method that directly renders table HTML to PDF
This uses the `weasyprint` package to generate the PDF, then
`pdfplumber` to determine the bounding box of table elements, and
finally `pypdf` to write the cropped PDF.
