+++ great_tables/gt.py
@@ -10,7 +10,7 @@ from great_tables._gt_data import GTData
 from great_tables._body import body_reassemble
 from great_tables._boxhead import cols_align, cols_label
 from great_tables._data_color import data_color
-from great_tables._export import as_raw_html, save, show
+from great_tables._export import as_raw_html, save, save_pdf, show
 from great_tables._formats import (
     fmt,
     fmt_bytes,
@@ -268,6 +268,7 @@ class GT(
     with_locale = with_locale
 
     save = save
+    save_pdf = save_pdf
     show = show
     as_raw_html = as_raw_html
 
