### 2024-10-19 649aa73
Ensure source notes are processed for LaTeX
