### 2024-10-19 a99a3d6
Ensure context is set in `_process_text()` call
