### 2024-10-19 7c95c41
Add `save_pdf()` to docs
