+++ docs/_quarto.yml
@@ -208,10 +208,12 @@ quartodoc:
     - title: Export
       desc: >
         There may come a day when you need to export a table to some specific format. A great method
-        for that is `save()`, which allows us to save the table as a standalone image file. You can
-        also get the table code as an HTML fragment with the `as_raw_html()` method.
+        for that is `save()`, which allows us to save the table as a standalone image file. To render
+        to a vector graphics PDF file, use `save_pdf()`. You can also get the table code as an HTML
+        fragment with the `as_raw_html()` method.
       contents:
         - GT.save
+        - GT.save_pdf
         - GT.show
         - GT.as_raw_html
     - title: Value formatting functions
