### 2024-10-20 f38f9cd
Reorganize LaTeX escaping functions and tests
