+++ great_tables/_text.py
@@ -89,3 +89,39 @@ def _latex_escape(text: str) -> str:
     text = re.sub(latex_escape_regex, lambda match: "\\" + match.group(), text)
 
     return text
+
+
+def escape_pattern_str_latex(pattern_str: str) -> str:
+
+    pattern = r"(\{[x0-9]+\})"
+
+    return process_string(pattern_str, pattern, _latex_escape)
+
+
+def process_string(string: str, pattern: str, func: Callable[[str], str]) -> str:
+    """
+    Process a string selectively based on a pattern.
+
+    This function splits a string based on a pattern and applies a function to elements that do not
+    match the pattern. The processed elements are then recombined to obtain a selectively processed
+    string.
+
+    Args:
+        string (str): The string to process.
+        pattern (str): The pattern to split the string by.
+        func (Callable[[str], str]): The function applied to elements that do not match the pattern.
+
+    Returns:
+        str: The selectively processed string.
+    """
+
+    # Split the string by the pattern
+    split_result = re.split(pattern, string)
+
+    # Apply the function to elements that do not match the pattern
+    processed_list = [func(part) if not re.match(pattern, part) else part for part in split_result]
+
+    # Recombine the list elements to obtain a selectively processed string
+    combined_str = "".join(processed_list)
+
+    return combined_str
