+++ great_tables/_utils.py
@@ -237,32 +237,3 @@ def seq_groups(seq: Iterable[str]) -> Generator[tuple[str, int], None, None]:
 
 def is_equal(x: Any, y: Any) -> bool:
     return x is not None and x == y
-
-
-def process_string(string: str, pattern: str, func: Callable[[str], str]) -> str:
-    """
-    Process a string selectively based on a pattern.
-
-    This function splits a string based on a pattern and applies a function to elements that do not
-    match the pattern. The processed elements are then recombined to obtain a selectively processed
-    string.
-
-    Args:
-        string (str): The string to process.
-        pattern (str): The pattern to split the string by.
-        func (Callable[[str], str]): The function applied to elements that do not match the pattern.
-
-    Returns:
-        str: The selectively processed string.
-    """
-
-    # Split the string by the pattern
-    split_result = re.split(pattern, string)
-
-    # Apply the function to elements that do not match the pattern
-    processed_list = [func(part) if not re.match(pattern, part) else part for part in split_result]
-
-    # Recombine the list elements to obtain a selectively processed string
-    combined_str = "".join(processed_list)
-
-    return combined_str
