### 2024-10-20 ece8d87
save_pdf(): Small clean-ups
