### 2024-10-21 6a738e8
Move _render_as_latex() out of gt.py
