### 2024-10-21 638980e
Add docstring
