### 2024-10-21 1778949
Use variable instead of hardcoded value
