### 2024-10-21 640933a
Modify docstring of `as_latex()`
