### 2024-10-21 f35ffc7
Ensure that title/subtitle undergo text processing
