### 2024-10-21 2b1f868
Remove unneeded comment
