### 2024-10-21 b1284b1
Create and use methods of Text/Md/Html
