### 2024-10-21 e401793
Add tests for generation of tbl fontsize stmt
