+++ tests/test_utils_render_latex.py
@@ -156,11 +156,37 @@ def test_create_width_dict_l_settings():
     assert width_dict["tbl_width"] is None
 
 
-def test_create_fontsize_statement_l():
-
-    gt_tbl = GT(exibble)
+def test_create_fontsize_statement_l(gt_tbl: GT):
 
     assert create_fontsize_statement_l(gt_tbl) == "\\fontsize{12.0pt}{14.4pt}\\selectfont\n"
+
+
+def test_create_fontsize_statement_l_px(gt_tbl: GT):
+
+    gt_tbl_new = gt_tbl.tab_options(table_font_size="11px")
+
+    assert create_fontsize_statement_l(gt_tbl_new) == "\\fontsize{8.2pt}{9.9pt}\\selectfont\n"
+
+
+def test_create_fontsize_statement_l_pct(gt_tbl: GT):
+
+    gt_tbl_new = gt_tbl.tab_options(table_font_size="50%")
+
+    assert create_fontsize_statement_l(gt_tbl_new) == "\\fontsize{6.0pt}{7.2pt}\\selectfont\n"
+
+
+def test_create_fontsize_statement_l_cm(gt_tbl: GT):
+
+    gt_tbl_new = gt_tbl.tab_options(table_font_size="0.6cm")
+
+    assert create_fontsize_statement_l(gt_tbl_new) == "\\fontsize{17.2pt}{20.7pt}\\selectfont\n"
+
+
+def test_create_fontsize_statement_l_unknown_unit(gt_tbl: GT):
+
+    gt_tbl_new = gt_tbl.tab_options(table_font_size="1span")
+
+    assert create_fontsize_statement_l(gt_tbl_new) == ""
 
 
 def test_create_fontsize_statement_l_settings():
