### 2024-10-21 50a8c4e
Add tests for deriving tbl width stmts
