### 2024-10-21 73616ea
Add test for creation of tbl fontsize stmt
