### 2024-10-26 #202 @Batalex
Hey @rich-iannone,
Sorry, it took me so long to get back to this review. Let me make it up to you by introducing the two rock stars who volunteered to review `great-tables`: @glemaitre and @cjbassin.

@glemaitre and @cjbassin, thank you so much or volunteering to review for pyOpenSci! I hope this experience will be as fulfilling as it was for me.

- [ ] @glemaitre survey completed.
- [x] @cjbassin survey completed.

Please get in touch with any questions or concerns! Your review is due: 18th nov.
---

Reviewers: @glemaitre,@cjbassin
Due date: 2024/11/18
