### 2024-10-31 #202 @cjbassin
Hi-I am  beginning the review but noticed that the original issue had a v0.9.0 version in the issue request, where as it looks like substantial improvements have been made in new releases since then, and its currently at v0.13.0. Should I install the older version or would the preference be to do a review on the current version?On Oct 26, 2024, at 12:34 PM, Alex Batisse ***@***.***> wrote:﻿
Hey @rich-iannone,
Sorry, it took me so long to get back to this review. Let me make it up to you by introducing the two rock stars who volunteered to review great-tables: @glemaitre and @cjbassin.
@glemaitre and @cjbassin, thank you so much or volunteering to review for pyOpenSci! I hope this experience will be as fulfilling as it was for me.
Please fill out our pre-review survey
Before beginning your review, please fill out our pre-review survey. This helps us improve all aspects of our review and better understand our community. No personal data will be shared from this survey - it will only be used in an aggregated format by our Executive Director to improve our processes and programs.

 @glemaitre survey completed.
 @cjbassin survey completed.

Here is the reviewers guide. This guide contains all of the steps and information needed to complete your review.
Here is the review template that you will need to fill out and submit

Please get in touch with any questions or concerns! Your review is due: 18th nov.
Reviewers: @***@***.***
Due date: 2024/11/18
