### 2024-11-11 #202 @CorinneBassin-SOI
---

Overall, I think this is a fantastic package that not only has clear documentation but has provided several ways for the community to be involved with discussion boards, discord, etc. My notes below are primarily to help users who might be less experienced.
A few Notes
- Installation instructions.
    - Installation using pip is clear and easy to find.
    - I also used `conda install conda-forge::great_tables` . If you are comfortable adding that as a possible secondary example on instructions, I always find it helpful to know what channel might be the best to install from.
    - Also, in order to run through all the examples using `from great_tables.data import islands` you have to have pandas installed. The error `ModuleNotFoundError: Currently, importing great_tables.data requires pandas. This will change in the future. Most of great_tables can be used without pandas.` that comes up is great, but it would be better to have it clear in the documentation at https://posit-dev.github.io/great-tables/get-started/#a-basic-table-using-great-tables that it needs to be installed to use this example data set at this point to lower the barrier of entry for users.
- Although you have a CONTRIBUTING.MD in the repository, it is not linked in the README or in the docs. Linking this along with where you have the Code of Conduct would make it easier to find.
- Examples
    - The basic example assumes you are not working at a command line. Make it clear either that examples should be run in a notebook environment or make the initial example include saving output to HTML in the documentation as proposed (https://github.com/posit-dev/great-tables/issues/409)

Item Not Checked in the Review Above:
[ ] Citation
    -  Governance and Engagement is clear, but I did not find a clear reference to CITATION in the README or DOCS
