### 2024-11-16 #202 @Batalex
Here is my own review, focusing on the code aspects. TBH, I mostly eyeballed than did a thorough line-by-line review considering how massive the codebase is. I hope you will get something out of it, though.

The review more or less follows the https://conventionalcomments.org/ structure, each category representing how strongly I feel about the point I raise. I want to make it clear to you that since I am the editor for this review, you do not have to follow it to get this package approved. Pick what you want!

# Praises

- praise (README.md): This README is mostly the perfect entry point to the project.
- praise (_types.py): What a great alternative to `typing.Self` for Python < 3.11!
- praise (_helper.py): The GFont helper is such a nice touch.
- praise (_formats.py, _formats_vals.py): Everything is so nicely laid out that it's a joy to follow

# Nitpicks

- nitpick (README.md): I would advise linking the `CONTRIBUTING.md` file somewhere in the README, probably near the Code of Conduct section. Currently, it is somewhat hidden in the `.github` folder.
- nitpick (CONTRIBUTING.md): We could add a link to https://stackoverflow.com/help/minimal-reproducible-example in the section `Filing issues`.
- nitpick (pyrightconfig.json): This is purely a personal preference, but I like having my pyright configuration in my `pyproject.toml` file. I am not expecting a change, I just wanted to let you know about this possibility.

# Suggestions

- suggestion (CONTRIBUTING.md): I believe we can ease the onboarding of new contributors and reduce the burden of maintenance by adding a little more content to this file.
 Namely, we can explicitly list the requirements for a submission to be considered (e.g. code style, formatting, having tests covering the new feature/preventing a regression on a bug fix) and the *make* command associated. In addition, we can also include the command to create a virtual environment with the development dependencies and guide new contributors to using pre-commit.
 - suggestion (_text.py): I wonder if we could use `hasattr(x, "to_html")` instead of checking for a `UnitStr` instance. That way we would 1. get rid of the circular import and 2. enable other types. I totally understand if you want to keep full control of what is rendered, though.
- suggestion (_text.py): Maybe we could make `Text` a subclass of `str` instead of a single attribute dataclass. We would get rid of quite a few types checks errors.
- suggestion (general): `great-tables`' code base is massive, so it's not a surprise that there is a lot going on, maintenance-wise. I'd like to share a tip I've used for a few years to manage my TODOs and FIXMEs throughout code bases. I use the syntax `# <CATEGORY>(<scope>): <comment>` (`FIXME`, respectively)(you may notice that it's the same "conventional" syntax that I use in this very same review). The idea is that you can quickly extract all the TODOs related to a specific subject. For instance, in `_utils.py`, in the `pairwise` function, the comment about Python 3.10 could be `# TODO(py310): Remove because part of stdlib`. With this one trick, you can easily track down all TODOs and FIXMEs related to Python 3.10.

# Issues

- issue (general): We have numerous unneeded imports throughout the files.
I recommend setting up a linter to track down unnecessary imports to reduce the boilerplate and make the format more consistent (for instance in `_helpers.py` where we have two different ways to import stuff from `_text.py`). `isort` is quite [easy to setup with black](https://pycqa.github.io/isort/docs/configuration/black_compatibility.html). `ruff` is another alternative which plays nice with `black` out of the box, though this discussion would then broaden the discussion to replacing `flake8` as well. Both solutions would also require removing comments from "imports blocs", something that I saw quite often in the code base.
- issue (general): There are a fair number of files with just one pass-through function. Is this because you want to keep the codebase as close as possible to `gt`'s, or is it just a refactoring waiting to be done?
- issue (_boxhead.py, _formats.py, _modify_rows.py, _spanners.py, _utils_nanoplots.py, others): I can spot a few places where we could move some import statements to the top of the file instead of being in functions. AFAICT there is no circular import issue there, and it does not seem to be for performance reasons either.
- issue (_boxhead.py):  The type annotation for `cols_align` is missing `| None` for the `columns` parameter.
- issue (_utils_nanoplots.py): I am a little on the fence about this file. It seems a little out of place, and I would not be surprised if this was extracted in its own package, just like `htmltools`. It has a LOT of content, its scope is way bigger than other files: data validation, statistic computation, and the actual SVG build. The thousand lines long function seems high maintenance as well.
