### 2024-12-02 posit-dev/great-tables#527 @rich-iannone
[pyOpenSci review] Link to the CONTRIBUTING.md in README.md
From the pyOpenSci reviewer comments:

Although you have a CONTRIBUTING.MD in the repository, it is not linked in the README or in the docs. Linking this along with where you have the Code of Conduct would make it easier to find.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2468592164

and also:

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
