### 2024-12-02 posit-dev/great-tables#528 @rich-iannone
[pyOpenSci review] Provide conda install instructions in README.md
From the pyOpenSci review comment:

I also used conda install conda-forge::great_tables . If you are comfortable adding that as a possible secondary example on instructions, I always find it helpful to know what channel might be the best to install from.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2468592164
