### 2024-12-02 posit-dev/great-tables#529 @rich-iannone
[pyOpenSci review] Add note about required use of Pandas for datasets
From the pyOpenSci review comment:

Also, in order to run through all the examples using from great_tables.data import islands you have to have pandas installed. The error ModuleNotFoundError: Currently, importing great_tables.data requires pandas. This will change in the future. Most of great_tables can be used without pandas. that comes up is great, but it would be better to have it clear in the documentation at https://posit-dev.github.io/great-tables/get-started/#a-basic-table-using-great-tables that it needs to be installed to use this example data set at this point to lower the barrier of entry for users.

So either remove this dependency or make note of it in the documentation.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2468592164
