### 2024-12-02 posit-dev/great-tables#530 @rich-iannone
[pyOpenSci review] Make clear in docs that examples to be run in notebook environment (not command line)
From the pyOpenSci comment:

The basic example assumes you are not working at a command line. Make it clear either that examples should be run in a notebook environment or make the initial example include saving output to HTML in the documentation as proposed

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2468592164
