### 2024-12-02 posit-dev/great-tables#531 @rich-iannone
[pyOpenSci review] Add a clear reference to CITATION in the README or DOCS
From the pyOpenSci comment:

Governance and Engagement is clear, but I did not find a clear reference to CITATION in the README or DOCS

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2468592164
