### 2024-12-02 posit-dev/great-tables#532 @rich-iannone
[pyOpenSci review] Link to MRE explainer in SO in the Filing Issues section of CONTRIBUTING.md
From the pyOpenSci comment:

We could add a link to https://stackoverflow.com/help/minimal-reproducible-example in the section Filing issues

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
