### 2024-12-02 posit-dev/great-tables#533 @rich-iannone
[pyOpenSci review] Add pyright config to pyproject.toml
From the pyOpenSci comment:

(pyrightconfig.json): This is purely a personal preference, but I like having my pyright configuration in my pyproject.toml file. I am not expecting a change, I just wanted to let you know about this possibility.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
