### 2024-12-02 posit-dev/great-tables#534 @rich-iannone
[pyOpenSci review] Improve onboarding of new contributors by additional info in CONTRIBUTING.md
From the pyOpenSci comment:

suggestion (CONTRIBUTING.md): I believe we can ease the onboarding of new contributors and reduce the burden of maintenance by adding a little more content to this file.
Namely, we can explicitly list the requirements for a submission to be considered (e.g. code style, formatting, having tests covering the new feature/preventing a regression on a bug fix) and the make command associated. In addition, we can also include the command to create a virtual environment with the development dependencies and guide new contributors to using pre-commit.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
