### 2024-12-02 posit-dev/great-tables#535 @rich-iannone
[pyOpenSci review] use `hasattr(x, "to_html")` instead of checking for a `UnitStr` instance
From the pyOpenSci comment:

suggestion (_text.py): I wonder if we could use hasattr(x, "to_html") instead of checking for a UnitStr instance. That way we would 1. get rid of the circular import and 2. enable other types. I totally understand if you want to keep full control of what is rendered, though.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
