### 2024-12-02 posit-dev/great-tables#536 @rich-iannone
[pyOpenSci review] make `Text` a subclass of `str` instead of a single attribute dataclass
From the pyOpenSci comment:

suggestion (_text.py): Maybe we could make Text a subclass of str instead of a single attribute dataclass. We would get rid of quite a few types checks errors.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
