### 2024-12-02 posit-dev/great-tables#537 @rich-iannone
[pyOpenSci review] improve TODOs and FIXMEs with additional context
From the pyOpenSci comment:

suggestion (general): great-tables' code base is massive, so it's not a surprise that there is a lot going on, maintenance-wise. I'd like to share a tip I've used for a few years to manage my TODOs and FIXMEs throughout code bases. I use the syntax # <CATEGORY>(<scope>): <comment> (FIXME, respectively)(you may notice that it's the same "conventional" syntax that I use in this very same review). The idea is that you can quickly extract all the TODOs related to a specific subject. For instance, in _utils.py, in the pairwise function, the comment about Python 3.10 could be # TODO(py310): Remove because part of stdlib. With this one trick, you can easily track down all TODOs and FIXMEs related to Python 3.10.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
