### 2024-12-02 posit-dev/great-tables#538 @rich-iannone
[pyOpenSci review] Use linter to help remove unneeded imports across codebase
From the pyOpenSci comment:

issue (general): We have numerous unneeded imports throughout the files.
I recommend setting up a linter to track down unnecessary imports to reduce the boilerplate and make the format more consistent (for instance in _helpers.py where we have two different ways to import stuff from _text.py). isort is quite [easy to setup with black](https://pycqa.github.io/isort/docs/configuration/black_compatibility.html). ruff is another alternative which plays nice with black out of the box, though this discussion would then broaden the discussion to replacing flake8 as well. Both solutions would also require removing comments from "imports blocs", something that I saw quite often in the code base.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
