### 2024-12-02 posit-dev/great-tables#539 @rich-iannone
[pyOpenSci review] Consider consolidating files with only one function into others
From the pyOpenSci comment:

issue (general): There are a fair number of files with just one pass-through function. Is this because you want to keep the codebase as close as possible to gt's, or is it just a refactoring waiting to be done?

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
