### 2024-12-02 posit-dev/great-tables#540 @rich-iannone
[pyOpenSci review] remove imports from function scope wherever possible
From the pyOpenSci comment:

issue (_boxhead.py, _formats.py, _modify_rows.py, _spanners.py, _utils_nanoplots.py, others): I can spot a few places where we could move some import statements to the top of the file instead of being in functions. AFAICT there is no circular import issue there, and it does not seem to be for performance reasons either.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
