### 2024-12-02 posit-dev/great-tables#541 @rich-iannone
[pyOpenSci review] add missing type annotation in `cols_align()` method
From the pyOpenSci comment:

issue (_boxhead.py): The type annotation for cols_align is missing | None for the columns parameter.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
