### 2024-12-02 posit-dev/great-tables#542 @rich-iannone
[pyOpenSci review] develop plan on whether to put nanoplots code into separate library
From the pyOpenSci comment:

issue (_utils_nanoplots.py): I am a little on the fence about this file. It seems a little out of place, and I would not be surprised if this was extracted in its own package, just like htmltools. It has a LOT of content, its scope is way bigger than other files: data validation, statistic computation, and the actual SVG build. The thousand lines long function seems high maintenance as well.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2480804958
