### 2024-12-02 posit-dev/great-tables#543 @rich-iannone
[pyOpenSci review] in CONTRIBUTING.md file add install instructions for editable mode and document the [all] and [dev] installs
From the pyOpenSci comment:

In the CONTRIBUTING.md file, it could be useful to mention how to install the package in editable mode and document the different options, e.g. [all], [dev], etc.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2489584108
