### 2024-12-02 posit-dev/great-tables#544 @rich-iannone
[pyOpenSci review] put contributing guidelines on GT website
From the pyOpenSci comment:

I would find it easier if the contributing guidelines would be on the HTML website (thought that the current guideline mentions the README or CONTRIBUTING file).

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2489584108
