### 2024-12-02 posit-dev/great-tables#545 @rich-iannone
[pyOpenSci review] Better integrate notebook-based examples
From the pyOpenSci comment:

Some of the examples are linking to the GitHub repository with pre-rendered notebooks. I don't like it so much and I usually prefer a solution like sphinx-gallery allowing a nicer integration with the documentation and provide examples that can be downloaded.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2489584108
