### 2024-12-02 posit-dev/great-tables#546 @rich-iannone
[pyOpenSci review] README.md badge for CI status stuck on `"No Status"`
From the pyOpenSci comment:

At the time of the review, the badge for the CIs indicated "No Status". Maybe it should be checked (the CI tests are passing).

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2489584108
