### 2024-12-02 posit-dev/great-tables#547 @rich-iannone
[pyOpenSci review] exclude TYPE_CHECKING statement from coverage report
From the pyOpenSci review:

In the code, there is some code using the if TYPE_CHECKING statement. This could be excluded from the coverage report since this is False at runtime.

https://github.com/pyOpenSci/software-submission/issues/202#issuecomment-2489584108
