### 2024-12-05 0e7843e
feat: add `fmt_flag()` method (#523)
* Add flag table CSV to data dir

* Rename CSV file

* Create x_country_names.csv

* Add getter fns for country flags and names

* Add automatic changes from ruff integration

* Add the fmt_flag() method

* Remove second return statement

* Modify type annotation on FmtFlag height attr

* Specify arg names when using FmtFlag()

* Convert numeric height inputs to px str values

* Convert parsed country codes to uppercase form

* Allow for 2- and 3-letter codes; error otherwise

* Refactor dictionary lookup for country name

* Remove repeated statement

* Add
[…]
