+++ docs/_quarto.yml
@@ -135,6 +135,7 @@ quartodoc:
         - GT.fmt_markdown
         - GT.fmt_units
         - GT.fmt_image
+        - GT.fmt_flag
         - GT.fmt_nanoplot
         - GT.fmt
         - GT.data_color
