### 2024-12-09 posit-dev/great-tables#541 @rich-iannone
The `cols_align()`method is `columns: SelectExpr = None` and `SelectExpr` already allows for a `None` value as requested.

https://github.com/posit-dev/great-tables/blob/0e7843e18c573c182560d608d0b56c4861dcba55/great_tables/_boxhead.py#L159
