### 2024-12-09 posit-dev/great-tables#547 @rich-iannone
There's a possible global solution here: https://github.com/nedbat/coveragepy/issues/831#issuecomment-517778185
