### 2024-12-09 posit-dev/great-tables#536 @machow
We have changed this to now use an interface called BaseText, with only `.to_html()` and `.to_latex()` methods.

The interface now only requires these methods, and does not specify what data / attributes the class should hold. I think this gets back at what the reviewer noted -- with the twist that now our interface now doesn't have opinions on how an object is instantiated.
