### 2024-12-12 #202 @rich-iannone
@CorinneBassin-SOI @Batalex @glemaitre sorry for the delayed response here, but thank you all for the detailed reviews! As you can see just above, we taken all of the reviewer comments and created separate issues in the Great Tables repo.

We're getting pretty close to finishing up but might need a few more days yet. The current plan is to close all of these issues on the repo and provide a long-form response for each of the comments here (linking back to the issue). Thanks for your patience here!
