### 2024-12-19 #202 @Batalex
As far as I am concerned, I am happy with the changes you made following my comments. But I am not a reviewer, so it's up to you, @cjbassin and @glemaitre, to go through the very detailed changelog Richard made, and check (or not) the box `The author has responded to my review and made changes to my satisfaction. I recommend approving this package.`

Thank you all!
