### 2025-01-03 #202 @Batalex
🎉  `great-tables` has been approved by pyOpenSci! Thank you @rich-iannone  for submitting `great-tables` and many thanks to @cjbassin and @glemaitre for reviewing this package! 😸

As an early `gt` enthusiast, I am so proud of our collective work here. A few years ago, I began to work on my own Python alternative and though my interest in working on it dwindled long before the first ever public release, today I am honoured that PyOpenSci took some part in `great-tables`.

- [ ] Add the badge for pyOpenSci peer-review to the README.md of `great-tables`. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/202)`.

@rich-iannone

- [ ] The repository is missing the `paper.md` file.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉

---
