### 2025-01-06 #202 @rich-iannone
Thank you all for everything! I enjoyed the review process immensely and I think the package is on a better track because of it. @Batalex I had no idea you had worked on a table package! Funny how things go.

I'll get to crossing off items on the Author Wrap Up Tasks list. Plus, I will work on the JOSS paper. This is all very exciting and thanks again everyone here for guiding me through the process!
