### 2025-02-12 #202 @lwasser
hi friends! I'm crossing off checklist items here!!

✔ @rich-iannone, it looks like you did take our post-survey. thank you for yoru time in doing that!!
you are also already in our Slack!! If you are interested in submitting a blog ppost about great tables to the pyOpenSci blog, we welcome you to do so!!

Can you please let us know where you are with the JOSS submission? i just need to update labels accordingly so our status updates here!! Congratulations on joining the pyOpenSci growing ecosystem of tools 🚀. It's awesome to have great tables here!!
