### 2025-02-13 #202 @lwasser
ok great. Let's just use the GitHub release, given it's likely Zenodo wasn't set up until 0.15 ? Does that sound good to you @Batalex ? I just updated the issue!  i'm just going through and cleaning up issues, closing when we can, etc! 🚀
