### 2025-02-19 #202 @rich-iannone
@lwasser Sorry for the delay(s) here! I'm still planning to get the JOSS paper/submission completed (shooting for beginning of March for that one). Also, I am interested in writing a blog post for the pyOpenSci blog (same time frame for a draft of that).

I did create an archive for Great Tables in Zenodo but used the latest version at the time (v0.15.0). The archive is here: https://zenodo.org/records/14618357. I could also create archives for earlier versions (like v0.14.0 and lower). I unfortunately couldn't use the Zenodo GH integration because of permissions issues with the org. But I don't mind creating archives manually. Let me know if there is anything I should do immediately on this end!
