### 2025-02-21 #202 @lwasser
hi @rich-iannone 👋🏻 No worries at all!!
We absolutely welcome a blog post, so whenever you are ready for that and if you have any questions, please let me know.

[Here is an overview of how to submit one to pyOpenSci](https://www.pyopensci.org/handbook/organization/social-media-blog.html#blog-post-submission-process)

If you are open to creating an zenodo archive for the accepted version of this package (o.14.0) that would be super awesome and welcome!! If you do that, please respond here and we can update that issue!!

thank you so much!! ✨ I know that times are crazy and everyone's bandwidth is limited!
