### 2025-04-30 #202 @lwasser
hi there @rich-iannone i'm following up again on this - specifically whether you plan to submit to JOSS and if you are able to drop an archive of great tables at version 0.14 (accepted here) to zenodo. We can always close without JOSS if it's too much to do but the zenodo tag would be great to have to wrap things up here!! I hope that you are well (and that maybe i'll see you at scipy or pycon us!
