### 2025-04-30 #202 @rich-iannone
@lwasser I am so sorry for leaving this unresolved for so long. I just added the `v0.14.0` release in Zenodo: https://zenodo.org/records/15312934. Let me know if that looks okay.

I'm still up for preparing a JOSS paper and I'll resolve to get a draft ready by next week.
