### 2025-04-30 #202 @lwasser
No worries, @rich-iannone You don't need to apologize! It's a crazy time and you are a busy human!

Thank you for the zenodo drop!! 🚀 that's perfect. When you submit to JOSS, can you please link to this issue so they know it's already been reviewed and also we can keep tabs on things here.

Once JOSS is complete we'll add the doi here to the issue as well. 🙌🏻
