### 2025-07-29 #202 @lwasser
Ok friends. Let's do this. This issue has been open for a very long time. @rich-iannone there is no pressure to submit to JOSS. But i think for now, let's close this issue! Rich, if you do end up sending it through JOSS, can you please let us know (maybe on slack or ping me here) and we can update the issue. But the review itself here is complete!! Many thanks everyone for your work on this review!
